% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_abstract_rnn_cell}
\alias{layer_abstract_rnn_cell}
\title{Abstract object representing an RNN cell}
\usage{
layer_abstract_rnn_cell(...)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Abstract object representing an RNN cell
}
\details{
See \href{https://www.tensorflow.org/guide/keras/rnn}{the Keras RNN API guide}
for details about the usage of RNN API.

This is the base class for implementing RNN cells with custom behavior.

Every \code{RNNCell} must have the properties below and implement \code{call} with
the signature \verb{(output, next_state) = call(input, state)}.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
