% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_activity_regularization}
\alias{layer_activity_regularization}
\title{Layer that applies an update to the cost function based input activity}
\usage{
layer_activity_regularization(object, l1 = 0, l2 = 0, ...)
}
\arguments{
\item{l1}{L1 regularization factor (positive float).}

\item{l2}{L2 regularization factor (positive float).}

\item{...}{standard layer arguments.}
}
\description{
Layer that applies an update to the cost function based input activity
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
