% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_add}
\alias{layer_add}
\title{Layer that adds a list of inputs}
\usage{
layer_add(object, ...)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Layer that adds a list of inputs
}
\details{
It takes as input a list of tensors,
all of the same shape, and returns
a single tensor (also of the same shape).
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
