% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_center_crop}
\alias{layer_center_crop}
\title{A preprocessing layer which crops images}
\usage{
layer_center_crop(object, height, width, ...)
}
\arguments{
\item{height}{Integer, the height of the output shape.}

\item{width}{Integer, the width of the output shape.}

\item{...}{standard layer arguments.}
}
\description{
A preprocessing layer which crops images
}
\details{
This layers crops the central portion of the images to a target size. If an
image is smaller than the target size, it will be resized and cropped
so as to return the largest possible window in the image that matches
the target aspect ratio.

Input pixel values can be of any range (e.g. \verb{[0., 1.)} or \verb{[0, 255]}) and
of integer or floating point dtype.
By default, the layer will output floats.

For an overview and full list of preprocessing layers, see the preprocessing
\href{https://www.tensorflow.org/guide/keras/preprocessing_layers}{guide}.

Input shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., height, width, channels)}, in \code{"channels_last"} format.

Output shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., target_height, target_width, channels)}.

If the input height/width is even and the target height/width is odd (or
inversely), the input image is left-padded by 1 pixel.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
