% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_cropping_3d}
\alias{layer_cropping_3d}
\title{Cropping layer for 3D data (e.g. spatial or spatio-temporal)}
\usage{
layer_cropping_3d(
  object,
  cropping = list(list(1L, 1L), list(1L, 1L), list(1L, 1L)),
  data_format = NULL,
  ...
)
}
\arguments{
\item{cropping}{Int, or list of 3 ints, or list of 3 lists of 2 ints.
\itemize{
\item If int: the same symmetric cropping
is applied to depth, height, and width.
\item If list of 3 ints: interpreted as two different
symmetric cropping values for depth, height, and width:
\verb{(symmetric_dim1_crop, symmetric_dim2_crop, symmetric_dim3_crop)}.
\item If list of 3 lists of 2 ints: interpreted as
\verb{((left_dim1_crop, right_dim1_crop), (left_dim2_crop, right_dim2_crop), (left_dim3_crop, right_dim3_crop))}
}}

\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch_size, spatial_dim1, spatial_dim2, spatial_dim3, channels)}
while \code{channels_first} corresponds to inputs with shape
\verb{(batch_size, channels, spatial_dim1, spatial_dim2, spatial_dim3)}.
When unspecified, uses
\code{image_data_format} value found in your Keras config file at
\verb{~/.keras/keras.json} (if exists) else 'channels_last'.
Defaults to 'channels_last'.}

\item{...}{standard layer arguments.}
}
\description{
Cropping layer for 3D data (e.g. spatial or spatio-temporal)
}
\details{
Examples:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> input_shape = (2, 28, 28, 10, 3)
>>> x = np.arange(np.prod(input_shape)).reshape(input_shape)
>>> y = tf.keras.layers.Cropping3D(cropping=(2, 4, 2))(x)
>>> print(y.shape)
(2, 24, 20, 6, 3)
}\if{html}{\out{</div>}}

Input shape:
5D tensor with shape:
\itemize{
\item If \code{data_format} is \code{"channels_last"}:
\verb{(batch_size, first_axis_to_crop, second_axis_to_crop, third_axis_to_crop, depth)}
\item If \code{data_format} is \code{"channels_first"}:
\verb{(batch_size, depth, first_axis_to_crop, second_axis_to_crop, third_axis_to_crop)}
}

Output shape:
5D tensor with shape:
\itemize{
\item If \code{data_format} is \code{"channels_last"}:
\verb{(batch_size, first_cropped_axis, second_cropped_axis, third_cropped_axis, depth)}
\item If \code{data_format} is \code{"channels_first"}:
\verb{(batch_size, depth, first_cropped_axis, second_cropped_axis, third_cropped_axis)}
}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
