% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_discretization}
\alias{layer_discretization}
\title{A preprocessing layer which buckets continuous features by ranges}
\usage{
layer_discretization(
  object,
  bin_boundaries = NULL,
  num_bins = NULL,
  epsilon = 0.01,
  output_mode = "int",
  sparse = FALSE,
  ...
)
}
\arguments{
\item{bin_boundaries}{A list of bin boundaries. The leftmost and rightmost bins
will always extend to \code{-inf} and \code{inf}, so \verb{bin_boundaries=[0., 1., 2.]}
generates bins \verb{(-inf, 0.)}, \verb{[0., 1.)}, \verb{[1., 2.)}, and \verb{[2., +inf)}.
If this option is set, \code{adapt()} should not be called.}

\item{num_bins}{The integer number of bins to compute. If this option is set,
\code{adapt()} should be called to learn the bin boundaries.}

\item{epsilon}{Error tolerance, typically a small fraction close to zero (e.g.
0.01). Higher values of epsilon increase the quantile approximation, and
hence result in more unequal buckets, but could improve performance
and resource consumption.}

\item{output_mode}{Specification for the output of the layer. Values can be
\code{"int"}, \code{"one_hot"}, \code{"multi_hot"}, or
\code{"count"} configuring the layer as follows:
\itemize{
\item \code{"int"}: Return the discretized bin indices directly.
\item \code{"one_hot"}: Encodes each individual element in the input into an
array the same size as \code{num_bins}, containing a 1 at the input's bin
index. If the last dimension is size 1, will encode on that
dimension.  If the last dimension is not size 1, will append a new
dimension for the encoded output.
\item \code{"multi_hot"}: Encodes each sample in the input into a single array
the same size as \code{num_bins}, containing a 1 for each bin index
index present in the sample. Treats the last dimension as the sample
dimension, if input shape is \verb{(..., sample_length)}, output shape
will be \verb{(..., num_tokens)}.
\item \code{"count"}: As \code{"multi_hot"}, but the int array contains a count of
the number of times the bin index appeared in the sample.
Defaults to \code{"int"}.
}}

\item{sparse}{Boolean. Only applicable to \code{"one_hot"}, \code{"multi_hot"},
and \code{"count"} output modes. If TRUE, returns a \code{SparseTensor} instead of
a dense \code{Tensor}. Defaults to \code{FALSE}.}

\item{...}{standard layer arguments.}
}
\description{
A preprocessing layer which buckets continuous features by ranges
}
\details{
This layer will place each element of its input data into one of several
contiguous ranges and output an integer index indicating which range each
element was placed in.

For an overview and full list of preprocessing layers, see the preprocessing
\href{https://www.tensorflow.org/guide/keras/preprocessing_layers}{guide}.

Input shape:
Any \code{tf.Tensor} or \code{tf.RaggedTensor} of dimension 2 or higher.

Output shape:
Same as input shape.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
