% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_dot}
\alias{layer_dot}
\title{Layer that computes a dot product between samples in two tensors}
\usage{
layer_dot(inputs, ..., axes, normalize = FALSE)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Layer that computes a dot product between samples in two tensors
}
\details{
E.g. if applied to a list of two tensors \code{a} and \code{b} of shape
\verb{(batch_size, n)}, the output will be a tensor of shape \verb{(batch_size, 1)}
where each entry \code{i} will be the dot product between
\code{a[i]} and \code{b[i]}.

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> x = np.arange(10).reshape(1, 5, 2)
>>> print(x)
[[[0 1]
  [2 3]
  [4 5]
  [6 7]
  [8 9]]]
>>> y = np.arange(10, 20).reshape(1, 2, 5)
>>> print(y)
[[[10 11 12 13 14]
  [15 16 17 18 19]]]
>>> tf.keras.layers.Dot(axes=(1, 2))([x, y])
<tf.Tensor: shape=(1, 2, 2), dtype=int64, numpy=
array([[[260, 360],
        [320, 445]]])>
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> x1 = tf.keras.layers.Dense(8)(np.arange(10).reshape(5, 2))
>>> x2 = tf.keras.layers.Dense(8)(np.arange(10, 20).reshape(5, 2))
>>> dotted = tf.keras.layers.Dot(axes=1)([x1, x2])
>>> dotted.shape
TensorShape([5, 1])
}\if{html}{\out{</div>}}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
