% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_elu}
\alias{layer_elu}
\title{Exponential Linear Unit}
\usage{
layer_elu(object, alpha = 1, ...)
}
\arguments{
\item{alpha}{Scale for the negative factor.}

\item{...}{standard layer arguments.}
}
\description{
Exponential Linear Unit
}
\details{
It follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    f(x) =  alpha * (exp(x) - 1.) for x < 0
    f(x) = x for x >= 0
}\if{html}{\out{</div>}}

Input shape:
Arbitrary. Use the keyword argument \code{input_shape}
(list of integers, does not include the samples axis)
when using this layer as the first layer in a model.

Output shape:
Same shape as the input.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
