% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_flatten}
\alias{layer_flatten}
\title{Flattens the input. Does not affect the batch size}
\usage{
layer_flatten(object, data_format = NULL, ...)
}
\arguments{
\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch, ..., channels)} while \code{channels_first} corresponds to
inputs with shape \verb{(batch, channels, ...)}.
When unspecified, uses
\code{image_data_format} value found in your Keras config file at
\verb{~/.keras/keras.json} (if exists) else 'channels_last'.
Defaults to 'channels_last'.}

\item{...}{standard layer arguments.}
}
\description{
Flattens the input. Does not affect the batch size
}
\details{
Note: If inputs are shaped \verb{(batch,)} without a feature axis, then
flattening adds an extra channel dimension and output shape is \verb{(batch, 1)}.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
