% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_gaussian_dropout}
\alias{layer_gaussian_dropout}
\title{Apply multiplicative 1-centered Gaussian noise}
\usage{
layer_gaussian_dropout(object, rate, seed = NULL, ...)
}
\arguments{
\item{rate}{Float, drop probability (as with \code{Dropout}).
The multiplicative noise will have
standard deviation \code{sqrt(rate / (1 - rate))}.}

\item{seed}{Integer, optional random seed to enable deterministic behavior.}

\item{...}{standard layer arguments.}
}
\description{
As it is a regularization layer, it is only active at training time.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
