% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_gaussian_noise}
\alias{layer_gaussian_noise}
\title{Apply additive zero-centered Gaussian noise}
\usage{
layer_gaussian_noise(object, stddev, seed = NULL, ...)
}
\arguments{
\item{stddev}{Float, standard deviation of the noise distribution.}

\item{seed}{Integer, optional random seed to enable deterministic behavior.}

\item{...}{standard layer arguments.}
}
\description{
Apply additive zero-centered Gaussian noise
}
\details{
This is useful to mitigate overfitting
(you could see it as a form of random data augmentation).
Gaussian Noise (GS) is a natural choice as corruption process
for real valued inputs.

As it is a regularization layer, it is only active at training time.

Call arguments:
inputs: Input tensor (of any rank).
training: Python boolean indicating whether the layer should behave in
training mode (adding noise) or in inference mode (doing nothing).

Input shape:
Arbitrary. Use the keyword argument \code{input_shape}
(list of integers, does not include the samples axis)
when using this layer as the first layer in a model.

Output shape:
Same shape as input.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
