% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_global_average_pooling_2d}
\alias{layer_global_average_pooling_2d}
\title{Global average pooling operation for spatial data}
\usage{
layer_global_average_pooling_2d(object, ...)
}
\arguments{
\item{...}{standard layer arguments.}

\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch, height, width, channels)} while \code{channels_first}
corresponds to inputs with shape
\verb{(batch, channels, height, width)}.
When unspecified, uses \code{image_data_format} value found
in your Keras config file at \verb{~/.keras/keras.json}
(if exists) else 'channels_last'. Defaults to 'channels_last'.}

\item{keepdims}{A boolean, whether to keep the spatial dimensions or not.
If \code{keepdims} is \code{FALSE} (default), the rank of the tensor is reduced
for spatial dimensions.
If \code{keepdims} is \code{TRUE}, the spatial dimensions are retained with
length 1.
The behavior is the same as for \code{tf.reduce_mean} or \code{np.mean}.}
}
\description{
Global average pooling operation for spatial data
}
\details{
\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> input_shape = (2, 4, 5, 3)
>>> x = tf.random.normal(input_shape)
>>> y = tf.keras.layers.GlobalAveragePooling2D()(x)
>>> print(y.shape)
(2, 3)
}\if{html}{\out{</div>}}

Input shape:
\itemize{
\item If \code{data_format='channels_last'}:
4D tensor with shape \verb{(batch_size, rows, cols, channels)}.
\item If \code{data_format='channels_first'}:
4D tensor with shape \verb{(batch_size, channels, rows, cols)}.
}

Output shape:
\itemize{
\item If \code{keepdims}=FALSE:
2D tensor with shape \verb{(batch_size, channels)}.
\item If \code{keepdims}=TRUE:
\itemize{
\item If \code{data_format='channels_last'}:
4D tensor with shape \verb{(batch_size, 1, 1, channels)}
\item If \code{data_format='channels_first'}:
4D tensor with shape \verb{(batch_size, channels, 1, 1)}
}
}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
