% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_global_average_pooling_3d}
\alias{layer_global_average_pooling_3d}
\title{Global Average pooling operation for 3D data}
\usage{
layer_global_average_pooling_3d(object, ...)
}
\arguments{
\item{...}{standard layer arguments.}

\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch, spatial_dim1, spatial_dim2, spatial_dim3, channels)}
while \code{channels_first} corresponds to inputs with shape
\verb{(batch, channels, spatial_dim1, spatial_dim2, spatial_dim3)}.
When unspecified, uses
\code{image_data_format} value found in your Keras config file at
\verb{~/.keras/keras.json} (if exists) else 'channels_last'.
Defaults to 'channels_last'.}

\item{keepdims}{A boolean, whether to keep the spatial dimensions or not.
If \code{keepdims} is \code{FALSE} (default), the rank of the tensor is reduced
for spatial dimensions.
If \code{keepdims} is \code{TRUE}, the spatial dimensions are retained with
length 1.
The behavior is the same as for \code{tf.reduce_mean} or \code{np.mean}.}
}
\description{
Global Average pooling operation for 3D data
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
