% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_group_normalization}
\alias{layer_group_normalization}
\title{Group normalization layer}
\usage{
layer_group_normalization(
  object,
  groups = 32L,
  axis = -1L,
  epsilon = 0.001,
  center = TRUE,
  scale = TRUE,
  beta_initializer = "zeros",
  gamma_initializer = "ones",
  beta_regularizer = NULL,
  gamma_regularizer = NULL,
  beta_constraint = NULL,
  gamma_constraint = NULL,
  ...
)
}
\arguments{
\item{groups}{Integer, the number of groups for Group Normalization. Can be in
the range \link{1, N} where N is the input dimension. The input dimension
must be divisible by the number of groups. Defaults to \code{32}.}

\item{axis}{Integer or List/Tuple. The axis or axes to normalize across.
Typically, this is the features axis/axes. The left-out axes are
typically the batch axis/axes. \code{-1} is the last dimension in the
input. Defaults to \code{-1}.}

\item{epsilon}{Small float added to variance to avoid dividing by zero. Defaults
to 1e-3}

\item{center}{If TRUE, add offset of \code{beta} to normalized tensor. If FALSE,
\code{beta} is ignored. Defaults to \code{TRUE}.}

\item{scale}{If TRUE, multiply by \code{gamma}. If FALSE, \code{gamma} is not used.
When the next layer is linear (also e.g. \code{nn.relu}), this can be
disabled since the scaling will be done by the next layer.
Defaults to \code{TRUE}.}

\item{beta_initializer}{Initializer for the beta weight. Defaults to zeros.}

\item{gamma_initializer}{Initializer for the gamma weight. Defaults to ones.}

\item{beta_regularizer}{Optional regularizer for the beta weight. NULL by
default.}

\item{gamma_regularizer}{Optional regularizer for the gamma weight. NULL by
default.}

\item{beta_constraint}{Optional constraint for the beta weight. NULL by default.}

\item{gamma_constraint}{Optional constraint for the gamma weight. NULL by
default.  Input shape: Arbitrary. Use the keyword argument \code{input_shape}
(list of integers, does not include the samples axis) when using this
layer as the first layer in a model.  Output shape: Same shape as input.}

\item{...}{standard layer arguments.}
}
\description{
Group normalization layer
}
\details{
Group Normalization divides the channels into groups and computes
within each group the mean and variance for normalization.
Empirically, its accuracy is more stable than batch norm in a wide
range of small batch sizes, if learning rate is adjusted linearly
with batch sizes.

Relation to Layer Normalization:
If the number of groups is set to 1, then this operation becomes nearly
identical to Layer Normalization (see Layer Normalization docs for details).

Relation to Instance Normalization:
If the number of groups is set to the input dimension (number of groups is
equal to number of channels), then this operation becomes identical to
Instance Normalization.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
