% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_identity}
\alias{layer_identity}
\title{Identity layer}
\usage{
layer_identity(object, ...)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Identity layer
}
\details{
This layer should be used as a placeholder when no operation is to be
performed. The layer is argument insensitive, and returns its \code{inputs}
argument as output.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
