% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_input}
\alias{layer_input}
\title{\code{Input()} is used to instantiate a Keras tensor}
\usage{
layer_input(
  shape = NULL,
  batch_size = NULL,
  name = NULL,
  dtype = NULL,
  sparse = NULL,
  tensor = NULL,
  ragged = NULL,
  type_spec = NULL,
  ...
)
}
\arguments{
\item{shape}{A shape list (integers), not including the batch size.
For instance, \verb{shape=(32,)} indicates that the expected input
will be batches of 32-dimensional vectors. Elements of this list
can be NULL; 'NULL' elements represent dimensions where the shape is
not known.}

\item{batch_size}{optional static batch size (integer).}

\item{sparse}{A boolean specifying whether the placeholder to be created is
sparse. Only one of 'ragged' and 'sparse' can be TRUE. Note that,
if \code{sparse} is FALSE, sparse tensors can still be passed into the
input - they will be densified with a default value of 0.}

\item{tensor}{Optional existing tensor to wrap into the \code{Input} layer.
If set, the layer will use the \code{tf.TypeSpec} of this tensor rather
than creating a new placeholder tensor.}

\item{ragged}{A boolean specifying whether the placeholder to be created is
ragged. Only one of 'ragged' and 'sparse' can be TRUE. In this case,
values of 'NULL' in the 'shape' argument represent ragged
dimensions.  For more information about RaggedTensors, see
\href{https://www.tensorflow.org/guide/ragged_tensor}{this guide}.}

\item{type_spec}{A \code{tf.TypeSpec} object to create the input placeholder from.
When provided, all other args except name must be NULL.}

\item{...}{standard layer arguments.}

\item{**kwargs}{deprecated arguments support. Supports \code{batch_shape} and
\code{batch_input_shape}.}
}
\description{
\code{Input()} is used to instantiate a Keras tensor
}
\details{
A Keras tensor is a symbolic tensor-like object, which we augment with
certain attributes that allow us to build a Keras model just by knowing the
inputs and outputs of the model.

For instance, if \code{a}, \code{b} and \code{c} are Keras tensors,
it becomes possible to do:
\verb{model = Model(input=[a, b], output=c)}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
