% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_masking}
\alias{layer_masking}
\title{Masks a sequence by using a mask value to skip timesteps}
\usage{
layer_masking(object, mask_value = 0, ...)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Masks a sequence by using a mask value to skip timesteps
}
\details{
For each timestep in the input tensor (dimension #1 in the tensor),
if all values in the input tensor at that timestep
are equal to \code{mask_value}, then the timestep will be masked (skipped)
in all downstream layers (as long as they support masking).

If any downstream layer does not support masking yet receives such
an input mask, an exception will be raised.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
