% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_max_pooling_1d}
\alias{layer_max_pooling_1d}
\title{Max pooling operation for 1D temporal data}
\usage{
layer_max_pooling_1d(
  object,
  pool_size = 2L,
  strides = NULL,
  padding = "valid",
  data_format = "channels_last",
  ...
)

layer_max_pooling_1d(
  object,
  pool_size = 2L,
  strides = NULL,
  padding = "valid",
  data_format = "channels_last",
  ...
)
}
\arguments{
\item{pool_size}{Integer, size of the max pooling window.}

\item{strides}{Integer, or NULL. Specifies how much the pooling window moves
for each pooling step.
If NULL, it will default to \code{pool_size}.}

\item{padding}{One of \code{"valid"} or \code{"same"} (case-insensitive).
\code{"valid"} means no padding. \code{"same"} results in padding evenly to
the left/right or up/down of the input such that output has the same
height/width dimension as the input.}

\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch, steps, features)} while \code{channels_first}
corresponds to inputs with shape
\verb{(batch, features, steps)}.}

\item{...}{standard layer arguments.}
}
\description{
Max pooling operation for 1D temporal data

Max pooling operation for 1D temporal data
}
\details{
Downsamples the input representation by taking the maximum value over a
spatial window of size \code{pool_size}. The window is shifted by \code{strides}.  The
resulting output, when using the \code{"valid"} padding option, has a shape of:
\verb{output_shape = (input_shape - pool_size + 1) / strides)}

The resulting output shape when using the \code{"same"} padding option is:
\code{output_shape = input_shape / strides}

For example, for \code{strides=1} and \code{padding="valid"}:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> x = tf.constant([1., 2., 3., 4., 5.])
>>> x = tf.reshape(x, [1, 5, 1])
>>> max_pool_1d = tf.keras.layers.MaxPooling1D(pool_size=2,
...    strides=1, padding='valid')
>>> max_pool_1d(x)
<tf.Tensor: shape=(1, 4, 1), dtype=float32, numpy=
array([[[2.],
        [3.],
        [4.],
        [5.]]], dtype=float32)>
}\if{html}{\out{</div>}}

For example, for \code{strides=2} and \code{padding="valid"}:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> x = tf.constant([1., 2., 3., 4., 5.])
>>> x = tf.reshape(x, [1, 5, 1])
>>> max_pool_1d = tf.keras.layers.MaxPooling1D(pool_size=2,
...    strides=2, padding='valid')
>>> max_pool_1d(x)
<tf.Tensor: shape=(1, 2, 1), dtype=float32, numpy=
array([[[2.],
        [4.]]], dtype=float32)>
}\if{html}{\out{</div>}}

For example, for \code{strides=1} and \code{padding="same"}:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> x = tf.constant([1., 2., 3., 4., 5.])
>>> x = tf.reshape(x, [1, 5, 1])
>>> max_pool_1d = tf.keras.layers.MaxPooling1D(pool_size=2,
...    strides=1, padding='same')
>>> max_pool_1d(x)
<tf.Tensor: shape=(1, 5, 1), dtype=float32, numpy=
array([[[2.],
        [3.],
        [4.],
        [5.],
        [5.]]], dtype=float32)>
}\if{html}{\out{</div>}}

Downsamples the input representation by taking the maximum value over a
spatial window of size \code{pool_size}. The window is shifted by \code{strides}.  The
resulting output, when using the \code{"valid"} padding option, has a shape of:
\verb{output_shape = (input_shape - pool_size + 1) / strides)}

The resulting output shape when using the \code{"same"} padding option is:
\code{output_shape = input_shape / strides}

For example, for \code{strides=1} and \code{padding="valid"}:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> x = tf.constant([1., 2., 3., 4., 5.])
>>> x = tf.reshape(x, [1, 5, 1])
>>> max_pool_1d = tf.keras.layers.MaxPooling1D(pool_size=2,
...    strides=1, padding='valid')
>>> max_pool_1d(x)
<tf.Tensor: shape=(1, 4, 1), dtype=float32, numpy=
array([[[2.],
        [3.],
        [4.],
        [5.]]], dtype=float32)>
}\if{html}{\out{</div>}}

For example, for \code{strides=2} and \code{padding="valid"}:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> x = tf.constant([1., 2., 3., 4., 5.])
>>> x = tf.reshape(x, [1, 5, 1])
>>> max_pool_1d = tf.keras.layers.MaxPooling1D(pool_size=2,
...    strides=2, padding='valid')
>>> max_pool_1d(x)
<tf.Tensor: shape=(1, 2, 1), dtype=float32, numpy=
array([[[2.],
        [4.]]], dtype=float32)>
}\if{html}{\out{</div>}}

For example, for \code{strides=1} and \code{padding="same"}:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> x = tf.constant([1., 2., 3., 4., 5.])
>>> x = tf.reshape(x, [1, 5, 1])
>>> max_pool_1d = tf.keras.layers.MaxPooling1D(pool_size=2,
...    strides=1, padding='same')
>>> max_pool_1d(x)
<tf.Tensor: shape=(1, 5, 1), dtype=float32, numpy=
array([[[2.],
        [3.],
        [4.],
        [5.],
        [5.]]], dtype=float32)>
}\if{html}{\out{</div>}}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}

\itemize{
\item \url{https://keras.io/api/layers}
}
}
