% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_max_pooling_3d}
\alias{layer_max_pooling_3d}
\title{Max pooling operation for 3D data (spatial or spatio-temporal)}
\usage{
layer_max_pooling_3d(
  object,
  pool_size = list(2L, 2L, 2L),
  strides = NULL,
  padding = "valid",
  data_format = NULL,
  ...
)

layer_max_pooling_3d(
  object,
  pool_size = list(2L, 2L, 2L),
  strides = NULL,
  padding = "valid",
  data_format = NULL,
  ...
)
}
\arguments{
\item{pool_size}{Tuple of 3 integers,
factors by which to downscale (dim1, dim2, dim3).
\verb{(2, 2, 2)} will halve the size of the 3D input in each dimension.}

\item{strides}{list of 3 integers, or NULL. Strides values.}

\item{padding}{One of \code{"valid"} or \code{"same"} (case-insensitive).
\code{"valid"} means no padding. \code{"same"} results in padding evenly to
the left/right or up/down of the input such that output has the same
height/width dimension as the input.}

\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch, spatial_dim1, spatial_dim2, spatial_dim3, channels)}
while \code{channels_first} corresponds to inputs with shape
\verb{(batch, channels, spatial_dim1, spatial_dim2, spatial_dim3)}.
When unspecified, uses
\code{image_data_format} value found in your Keras config file at
\verb{~/.keras/keras.json} (if exists) else 'channels_last'.
Defaults to 'channels_last'.}

\item{...}{standard layer arguments.}
}
\description{
Max pooling operation for 3D data (spatial or spatio-temporal)

Max pooling operation for 3D data (spatial or spatio-temporal)
}
\details{
Downsamples the input along its spatial dimensions (depth, height, and
width) by taking the maximum value over an input window (of size defined by
\code{pool_size}) for each channel of the input.  The window is shifted by
\code{strides} along each dimension.

Downsamples the input along its spatial dimensions (depth, height, and
width) by taking the maximum value over an input window (of size defined by
\code{pool_size}) for each channel of the input.  The window is shifted by
\code{strides} along each dimension.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}

\itemize{
\item \url{https://keras.io/api/layers}
}
}
