% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_maximum}
\alias{layer_maximum}
\title{Layer that computes the maximum (element-wise) a list of inputs}
\usage{
layer_maximum(object, ...)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Layer that computes the maximum (element-wise) a list of inputs
}
\details{
It takes as input a list of tensors, all of the same shape, and returns
a single tensor (also of the same shape).

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> tf.keras.layers.Maximum()([np.arange(5).reshape(5, 1),
...                            np.arange(5, 10).reshape(5, 1)])
<tf.Tensor: shape=(5, 1), dtype=int64, numpy=
array([[5],
     [6],
     [7],
     [8],
     [9]])>
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> x1 = tf.keras.layers.Dense(8)(np.arange(10).reshape(5, 2))
>>> x2 = tf.keras.layers.Dense(8)(np.arange(10, 20).reshape(5, 2))
>>> maxed = tf.keras.layers.Maximum()([x1, x2])
>>> maxed.shape
TensorShape([5, 8])
}\if{html}{\out{</div>}}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
