% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_minimum}
\alias{layer_minimum}
\title{Layer that computes the minimum (element-wise) a list of inputs}
\usage{
layer_minimum(object, ...)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Layer that computes the minimum (element-wise) a list of inputs
}
\details{
It takes as input a list of tensors, all of the same shape, and returns
a single tensor (also of the same shape).

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> tf.keras.layers.Minimum()([np.arange(5).reshape(5, 1),
...                            np.arange(5, 10).reshape(5, 1)])
<tf.Tensor: shape=(5, 1), dtype=int64, numpy=
array([[0],
     [1],
     [2],
     [3],
     [4]])>
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> x1 = tf.keras.layers.Dense(8)(np.arange(10).reshape(5, 2))
>>> x2 = tf.keras.layers.Dense(8)(np.arange(10, 20).reshape(5, 2))
>>> minned = tf.keras.layers.Minimum()([x1, x2])
>>> minned.shape
TensorShape([5, 8])
}\if{html}{\out{</div>}}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
