% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_normalization}
\alias{layer_normalization}
\title{A preprocessing layer which normalizes continuous features}
\usage{
layer_normalization(
  object,
  axis = -1L,
  mean = NULL,
  variance = NULL,
  invert = FALSE,
  ...
)
}
\arguments{
\item{axis}{Integer, list of integers, or NULL. The axis or axes that should
have a separate mean and variance for each index in the shape. For
example, if shape is \verb{(NULL, 5)} and \code{axis=1}, the layer will track 5
separate mean and variance values for the last axis. If \code{axis} is set
to \code{NULL}, the layer will normalize all elements in the input by a
scalar mean and variance. When \code{-1} the last axis of the
input is assumed to be a feature dimension and is normalized per
index. Note that in the specific case of batched scalar inputs where
the only axis is the batch axis, the default will normalize each index
in the batch separately. In this case, consider passing \code{axis=NULL}.
Defaults to \code{-1}.}

\item{mean}{The mean value(s) to use during normalization. The passed value(s)
will be broadcast to the shape of the kept axes above; if the value(s)
cannot be broadcast, an error will be raised when this layer's
\code{build()} method is called.}

\item{variance}{The variance value(s) to use during normalization. The passed
value(s) will be broadcast to the shape of the kept axes above; if the
value(s) cannot be broadcast, an error will be raised when this
layer's \code{build()} method is called.}

\item{invert}{If TRUE, this layer will apply the inverse transformation
to its inputs: it would turn a normalized input back into its
original form.}

\item{...}{standard layer arguments.}
}
\description{
A preprocessing layer which normalizes continuous features
}
\details{
This layer will shift and scale inputs into a distribution centered around
0 with standard deviation 1. It accomplishes this by precomputing the mean
and variance of the data, and calling \code{(input - mean) / sqrt(var)} at
runtime.

The mean and variance values for the layer must be either supplied on
construction or learned via \code{adapt()}. \code{adapt()} will compute the mean and
variance of the data and store them as the layer's weights. \code{adapt()} should
be called before \code{fit()}, \code{evaluate()}, or \code{predict()}.

For an overview and full list of preprocessing layers, see the preprocessing
\href{https://www.tensorflow.org/guide/keras/preprocessing_layers}{guide}.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
