% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_permute}
\alias{layer_permute}
\title{Permutes the dimensions of the input according to a given pattern}
\usage{
layer_permute(object, dims, ...)
}
\arguments{
\item{dims}{Tuple of integers. Permutation pattern does not include the
samples dimension. Indexing starts at 1.
For instance, \verb{(2, 1)} permutes the first and second dimensions
of the input.}

\item{...}{standard layer arguments.}
}
\description{
Permutes the dimensions of the input according to a given pattern
}
\details{
Useful e.g. connecting RNNs and convnets.

Example:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{model = Sequential()
model.add(Permute((2, 1), input_shape=(10, 64)))
# now: model.output_shape == (NULL, 64, 10)
# note: `NULL` is the batch dimension
}\if{html}{\out{</div>}}

Input shape:
Arbitrary. Use the keyword argument \code{input_shape}
(list of integers, does not include the samples axis)
when using this layer as the first layer in a model.

Output shape:
Same as the input shape, but with the dimensions re-ordered according
to the specified pattern.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
