% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_random_contrast}
\alias{layer_random_contrast}
\title{A preprocessing layer which randomly adjusts contrast during training}
\usage{
layer_random_contrast(object, factor, seed = NULL, ...)
}
\arguments{
\item{factor}{a positive float represented as fraction of value, or a list of
size 2 representing lower and upper bound.
When represented as a single float, lower = upper.
The contrast factor will be randomly picked between
\verb{[1.0 - lower, 1.0 + upper]}. For any pixel x in the channel,
the output will be \code{(x - mean) * factor + mean}
where \code{mean} is the mean value of the channel.}

\item{seed}{Integer. Used to create a random seed.}

\item{...}{standard layer arguments.}
}
\description{
A preprocessing layer which randomly adjusts contrast during training
}
\details{
This layer will randomly adjust the contrast of an image or images
by a random factor. Contrast is adjusted independently
for each channel of each image during training.

For each channel, this layer computes the mean of the image pixels in the
channel and then adjusts each component \code{x} of each pixel to
\code{(x - mean) * contrast_factor + mean}.

Input pixel values can be of any range (e.g. \verb{[0., 1.)} or \verb{[0, 255]}) and
in integer or floating point dtype.
By default, the layer will output floats.
The output value will be clipped to the range \verb{[0, 255]}, the valid
range of RGB colors.

For an overview and full list of preprocessing layers, see the preprocessing
\href{https://www.tensorflow.org/guide/keras/preprocessing_layers}{guide}.

Input shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., height, width, channels)}, in \code{"channels_last"} format.

Output shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., height, width, channels)}, in \code{"channels_last"} format.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
