% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_random_crop}
\alias{layer_random_crop}
\title{A preprocessing layer which randomly crops images during training}
\usage{
layer_random_crop(object, height, width, seed = NULL, ...)
}
\arguments{
\item{height}{Integer, the height of the output shape.}

\item{width}{Integer, the width of the output shape.}

\item{seed}{Integer. Used to create a random seed.}

\item{...}{standard layer arguments.}
}
\description{
A preprocessing layer which randomly crops images during training
}
\details{
During training, this layer will randomly choose a location to crop images
down to a target size. The layer will crop all the images in the same batch
to the same cropping location.

At inference time, and during training if an input image is smaller than the
target size, the input will be resized and cropped so as to return the
largest possible window in the image that matches the target aspect ratio.
If you need to apply random cropping at inference time, set \code{training} to
TRUE when calling the layer.

Input pixel values can be of any range (e.g. \verb{[0., 1.)} or \verb{[0, 255]}) and
of integer or floating point dtype. By default, the layer will output
floats.

For an overview and full list of preprocessing layers, see the preprocessing
\href{https://www.tensorflow.org/guide/keras/preprocessing_layers}{guide}.

Input shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., height, width, channels)}, in \code{"channels_last"} format.

Output shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., target_height, target_width, channels)}.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
