% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_random_flip}
\alias{layer_random_flip}
\title{A preprocessing layer which randomly flips images during training}
\usage{
layer_random_flip(object, mode = "horizontal_and_vertical", seed = NULL, ...)
}
\arguments{
\item{mode}{String indicating which flip mode to use. Can be \code{"horizontal"},
\code{"vertical"}, or \code{"horizontal_and_vertical"}. \code{"horizontal"} is a
left-right flip and \code{"vertical"} is a top-bottom flip. Defaults to
\code{"horizontal_and_vertical"}}

\item{seed}{Integer. Used to create a random seed.}

\item{...}{standard layer arguments.}
}
\description{
A preprocessing layer which randomly flips images during training
}
\details{
This layer will flip the images horizontally and or vertically based on the
\code{mode} attribute. During inference time, the output will be identical to
input. Call the layer with \code{training=TRUE} to flip the input.

Input pixel values can be of any range (e.g. \verb{[0., 1.)} or \verb{[0, 255]}) and
of integer or floating point dtype.
By default, the layer will output floats.

For an overview and full list of preprocessing layers, see the preprocessing
\href{https://www.tensorflow.org/guide/keras/preprocessing_layers}{guide}.

Input shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., height, width, channels)}, in \code{"channels_last"} format.

Output shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., height, width, channels)}, in \code{"channels_last"} format.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
