% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_random_height}
\alias{layer_random_height}
\title{A preprocessing layer which randomly varies image height during training}
\usage{
layer_random_height(
  object,
  factor,
  interpolation = "bilinear",
  seed = NULL,
  ...
)
}
\arguments{
\item{factor}{A positive float (fraction of original height),
or a list of size 2 representing lower and upper bound
for resizing vertically. When represented as a single float,
this value is used for both the upper and
lower bound. For instance, \verb{factor=(0.2, 0.3)} results
in an output with
height changed by a random amount in the range \verb{[20\%, 30\%]}.
\verb{factor=(-0.2, 0.3)} results in an output with height
changed by a random amount in the range \verb{[-20\%, +30\%]}.
\code{factor=0.2} results in an output with
height changed by a random amount in the range \verb{[-20\%, +20\%]}.}

\item{interpolation}{String, the interpolation method.
Supports \code{"bilinear"}, \code{"nearest"}, \code{"bicubic"}, \code{"area"},
\code{"lanczos3"}, \code{"lanczos5"}, \code{"gaussian"}, \code{"mitchellcubic"}.
Defaults to \code{"bilinear"}.}

\item{seed}{Integer. Used to create a random seed.}

\item{...}{standard layer arguments.}
}
\description{
A preprocessing layer which randomly varies image height during training
}
\details{
This layer adjusts the height of a batch of images by a random factor.
The input should be a 3D (unbatched) or 4D (batched) tensor in the
\code{"channels_last"} image data format. Input pixel values can be of any range
(e.g. \verb{[0., 1.)} or \verb{[0, 255]}) and of integer or floating point dtype. By
default, the layer will output floats.

By default, this layer is inactive during inference.

For an overview and full list of preprocessing layers, see the preprocessing
\href{https://www.tensorflow.org/guide/keras/preprocessing_layers}{guide}.

Input shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., height, width, channels)}, in \code{"channels_last"} format.

Output shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., random_height, width, channels)}.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
