% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_rescaling}
\alias{layer_rescaling}
\title{A preprocessing layer which rescales input values to a new range}
\usage{
layer_rescaling(object, scale, offset = 0, ...)
}
\arguments{
\item{scale}{Float, the scale to apply to the inputs.}

\item{offset}{Float, the offset to apply to the inputs.}

\item{...}{standard layer arguments.}
}
\description{
A preprocessing layer which rescales input values to a new range
}
\details{
This layer rescales every value of an input (often an image) by multiplying
by \code{scale} and adding \code{offset}.

For instance:
\enumerate{
\item To rescale an input in the \verb{[0, 255]} range
to be in the \verb{[0, 1]} range, you would pass \code{scale=1./255}.
\item To rescale an input in the \verb{[0, 255]} range to be in the \verb{[-1, 1]} range,
you would pass \verb{scale=1./127.5, offset=-1}.
}

The rescaling is applied both during training and inference. Inputs can be
of integer or floating point dtype, and by default the layer will output
floats.

For an overview and full list of preprocessing layers, see the preprocessing
\href{https://www.tensorflow.org/guide/keras/preprocessing_layers}{guide}.

Input shape:
Arbitrary.

Output shape:
Same as input.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
