% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_reshape}
\alias{layer_reshape}
\title{Layer that reshapes inputs into the given shape}
\usage{
layer_reshape(object, target_shape, ...)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Layer that reshapes inputs into the given shape
}
\details{
Input shape:
Arbitrary, although all dimensions in the input shape must be known/fixed.
Use the keyword argument \code{input_shape} (list of integers, does not
include the samples/batch size axis) when using this layer as the first
layer in a model.

Output shape:
\verb{(batch_size,) + target_shape}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
