% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_resizing}
\alias{layer_resizing}
\title{A preprocessing layer which resizes images}
\usage{
layer_resizing(
  object,
  height,
  width,
  interpolation = "bilinear",
  crop_to_aspect_ratio = FALSE,
  ...
)
}
\arguments{
\item{height}{Integer, the height of the output shape.}

\item{width}{Integer, the width of the output shape.}

\item{interpolation}{String, the interpolation method.
Supports \code{"bilinear"}, \code{"nearest"}, \code{"bicubic"}, \code{"area"},
\code{"lanczos3"}, \code{"lanczos5"}, \code{"gaussian"}, \code{"mitchellcubic"}.
Defaults to \code{"bilinear"}.}

\item{crop_to_aspect_ratio}{If TRUE, resize the images without aspect
ratio distortion. When the original aspect ratio differs
from the target aspect ratio, the output image will be
cropped so as to return the
largest possible window in the image (of size \verb{(height, width)})
that matches the target aspect ratio. By default
(\code{crop_to_aspect_ratio=FALSE}), aspect ratio may not be preserved.}

\item{...}{standard layer arguments.}
}
\description{
A preprocessing layer which resizes images
}
\details{
This layer resizes an image input to a target height and width. The input
should be a 4D (batched) or 3D (unbatched) tensor in \code{"channels_last"}
format. Input pixel values can be of any range
(e.g. \verb{[0., 1.)} or \verb{[0, 255]}) and of integer or floating point dtype.
By default, the layer will output floats.

This layer can be called on tf.RaggedTensor batches of input images of
distinct sizes, and will resize the outputs to dense tensors of uniform
size.

For an overview and full list of preprocessing layers, see the preprocessing
\href{https://www.tensorflow.org/guide/keras/preprocessing_layers}{guide}.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
