% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_separable_conv_1d}
\alias{layer_separable_conv_1d}
\title{Depthwise separable 1D convolution}
\usage{
layer_separable_conv_1d(
  object,
  filters,
  kernel_size,
  strides = 1L,
  padding = "valid",
  data_format = NULL,
  dilation_rate = 1L,
  depth_multiplier = 1L,
  activation = NULL,
  use_bias = TRUE,
  depthwise_initializer = "glorot_uniform",
  pointwise_initializer = "glorot_uniform",
  bias_initializer = "zeros",
  depthwise_regularizer = NULL,
  pointwise_regularizer = NULL,
  bias_regularizer = NULL,
  activity_regularizer = NULL,
  depthwise_constraint = NULL,
  pointwise_constraint = NULL,
  bias_constraint = NULL,
  ...
)

layer_separable_conv_1d(
  object,
  filters,
  kernel_size,
  strides = 1L,
  padding = "valid",
  data_format = NULL,
  dilation_rate = 1L,
  depth_multiplier = 1L,
  activation = NULL,
  use_bias = TRUE,
  depthwise_initializer = "glorot_uniform",
  pointwise_initializer = "glorot_uniform",
  bias_initializer = "zeros",
  depthwise_regularizer = NULL,
  pointwise_regularizer = NULL,
  bias_regularizer = NULL,
  activity_regularizer = NULL,
  depthwise_constraint = NULL,
  pointwise_constraint = NULL,
  bias_constraint = NULL,
  ...
)
}
\arguments{
\item{filters}{Integer, the dimensionality of the output space (i.e. the number
of filters in the convolution).}

\item{kernel_size}{A single integer specifying the spatial
dimensions of the filters.}

\item{strides}{A single integer specifying the strides
of the convolution.
Specifying any \code{stride} value != 1 is incompatible with specifying
any \code{dilation_rate} value != 1.}

\item{padding}{One of \code{"valid"}, \code{"same"}, or \code{"causal"} (case-insensitive).
\code{"valid"} means no padding. \code{"same"} results in padding with zeros
evenly to the left/right or up/down of the input such that output has
the same height/width dimension as the input. \code{"causal"} results in
causal (dilated) convolutions, e.g. \code{output[t]} does not depend on
\verb{input[t+1:]}.}

\item{data_format}{A string, one of \code{channels_last} (default) or
\code{channels_first}.  The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch_size, length, channels)} while \code{channels_first} corresponds to
inputs with shape \verb{(batch_size, channels, length)}.}

\item{dilation_rate}{A single integer, specifying
the dilation rate to use for dilated convolution.}

\item{depth_multiplier}{The number of depthwise convolution output channels for
each input channel. The total number of depthwise convolution output
channels will be equal to \code{num_filters_in * depth_multiplier}.}

\item{activation}{Activation function to use.
If you don't specify anything, no activation is applied
(see \code{keras.activations}).}

\item{use_bias}{Boolean, whether the layer uses a bias.}

\item{depthwise_initializer}{An initializer for the depthwise convolution kernel
(see \code{keras.initializers}). If NULL, then the default initializer
('glorot_uniform') will be used.}

\item{pointwise_initializer}{An initializer for the pointwise convolution kernel
(see \code{keras.initializers}). If NULL, then the default initializer
('glorot_uniform') will be used.}

\item{bias_initializer}{An initializer for the bias vector. If NULL, the default
initializer ('zeros') will be used (see \code{keras.initializers}).}

\item{depthwise_regularizer}{Optional regularizer for the depthwise
convolution kernel (see \code{keras.regularizers}).}

\item{pointwise_regularizer}{Optional regularizer for the pointwise
convolution kernel (see \code{keras.regularizers}).}

\item{bias_regularizer}{Optional regularizer for the bias vector
(see \code{keras.regularizers}).}

\item{activity_regularizer}{Optional regularizer function for the output
(see \code{keras.regularizers}).}

\item{depthwise_constraint}{Optional projection function to be applied to the
depthwise kernel after being updated by an \code{Optimizer} (e.g. used for
norm constraints or value constraints for layer weights). The function
must take as input the unprojected variable and must return the
projected variable (which must have the same shape). Constraints are
not safe to use when doing asynchronous distributed training
(see \code{keras.constraints}).}

\item{pointwise_constraint}{Optional projection function to be applied to the
pointwise kernel after being updated by an \code{Optimizer}
(see \code{keras.constraints}).}

\item{bias_constraint}{Optional projection function to be applied to the
bias after being updated by an \code{Optimizer}
(see \code{keras.constraints}).}

\item{...}{standard layer arguments.}

\item{trainable}{Boolean, if \code{TRUE} the weights of this layer will be marked as
trainable (and listed in \code{layer.trainable_weights}).}
}
\description{
Depthwise separable 1D convolution

Depthwise separable 1D convolution
}
\details{
This layer performs a depthwise convolution that acts separately on
channels, followed by a pointwise convolution that mixes channels.
If \code{use_bias} is TRUE and a bias initializer is provided,
it adds a bias vector to the output.
It then optionally applies an activation function to produce the final
output.

This layer performs a depthwise convolution that acts separately on
channels, followed by a pointwise convolution that mixes channels.
If \code{use_bias} is TRUE and a bias initializer is provided,
it adds a bias vector to the output.
It then optionally applies an activation function to produce the final
output.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}

\itemize{
\item \url{https://keras.io/api/layers}
}
}
