% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_softmax}
\alias{layer_softmax}
\title{Softmax activation function}
\usage{
layer_softmax(object, axis = -1L, ...)
}
\arguments{
\item{axis}{Integer, or list of Integers, axis along which the softmax
normalization is applied.}

\item{...}{standard layer arguments.}
}
\description{
Softmax activation function
}
\details{
Example without mask:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> inp = np.asarray([[1., 2., 1.]])
>>> layer = tf.keras.layers.Softmax()
>>> layer(inp).numpy()
array([[0.21194157, 0.5761169 , 0.21194157]], dtype=float32)
>>> mask = np.asarray([[TRUE, FALSE, TRUE]], dtype=bool)
>>> layer(inp, mask).numpy()
array([[0.5, 0. , 0.5]], dtype=float32)
}\if{html}{\out{</div>}}

Input shape:
Arbitrary. Use the keyword argument \code{input_shape}
(list of integers, does not include the samples axis)
when using this layer as the first layer in a model.

Output shape:
Same shape as the input.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
