% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_spatial_dropout_1d}
\alias{layer_spatial_dropout_1d}
\title{Spatial 1D version of Dropout}
\usage{
layer_spatial_dropout_1d(object, rate, ...)
}
\arguments{
\item{rate}{Float between 0 and 1. Fraction of the input units to drop.}

\item{...}{standard layer arguments.}
}
\description{
Spatial 1D version of Dropout
}
\details{
This version performs the same function as Dropout, however, it drops
entire 1D feature maps instead of individual elements. If adjacent frames
within feature maps are strongly correlated (as is normally the case in
early convolution layers) then regular dropout will not regularize the
activations and will otherwise just result in an effective learning rate
decrease. In this case, SpatialDropout1D will help promote independence
between feature maps and should be used instead.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
