% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_stacked_rnn_cells}
\alias{layer_stacked_rnn_cells}
\title{Wrapper allowing a stack of RNN cells to behave as a single cell}
\usage{
layer_stacked_rnn_cells(object, cells, ...)
}
\arguments{
\item{cells}{List of RNN cell instances.}

\item{...}{standard layer arguments.}
}
\description{
Used to implement efficient stacked RNNs.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
