% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_subtract}
\alias{layer_subtract}
\title{Layer that subtracts two inputs}
\usage{
layer_subtract(object, ...)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Layer that subtracts two inputs
}
\details{
It takes as input a list of tensors of size 2, both of the same shape, and
returns a single tensor, (inputs\link{0} - inputs\link{1}), also of the same shape.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
