% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_unit_normalization}
\alias{layer_unit_normalization}
\title{Unit normalization layer}
\usage{
layer_unit_normalization(object, axis = -1L, ...)
}
\arguments{
\item{axis}{Integer or list. The axis or axes to normalize across.
Typically, this is the features axis or axes. The left-out axes are
typically the batch axis or axes. \code{-1} is the last dimension
in the input. Defaults to \code{-1}.}

\item{...}{standard layer arguments.}
}
\description{
Unit normalization layer
}
\details{
Normalize a batch of inputs so that each input in the batch has a L2 norm
equal to 1 (across the axes specified in \code{axis}).

Example:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> data = tf.constant(np.arange(6).reshape(2, 3), dtype=tf.float32)
>>> normalized_data = tf.keras.layers.UnitNormalization()(data)
>>> print(tf.reduce_sum(normalized_data[0, :] ** 2).numpy())
1.0
}\if{html}{\out{</div>}}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
