% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_up_sampling_2d}
\alias{layer_up_sampling_2d}
\title{Upsampling layer for 2D inputs}
\usage{
layer_up_sampling_2d(
  object,
  size = list(2L, 2L),
  data_format = NULL,
  interpolation = "nearest",
  ...
)
}
\arguments{
\item{size}{Int, or list of 2 integers.
The upsampling factors for rows and columns.}

\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch_size, height, width, channels)} while \code{channels_first}
corresponds to inputs with shape
\verb{(batch_size, channels, height, width)}.
When unspecified, uses
\code{image_data_format} value found in your Keras config file at
\verb{~/.keras/keras.json} (if exists) else 'channels_last'.
Defaults to 'channels_last'.}

\item{interpolation}{A string, one of \code{"area"}, \code{"bicubic"}, \code{"bilinear"},
\code{"gaussian"}, \code{"lanczos3"}, \code{"lanczos5"}, \code{"mitchellcubic"},
\code{"nearest"}.}

\item{...}{standard layer arguments.}
}
\description{
Upsampling layer for 2D inputs
}
\details{
Repeats the rows and columns of the data
by \code{size[0]} and \code{size[1]} respectively.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
