% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_upsampling_3d}
\alias{layer_upsampling_3d}
\title{Upsampling layer for 3D inputs}
\usage{
layer_upsampling_3d(object, size = list(2L, 2L, 2L), data_format = NULL, ...)
}
\arguments{
\item{size}{Int, or list of 3 integers.
The upsampling factors for dim1, dim2 and dim3.}

\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch_size, spatial_dim1, spatial_dim2, spatial_dim3, channels)}
while \code{channels_first} corresponds to inputs with shape
\verb{(batch_size, channels, spatial_dim1, spatial_dim2, spatial_dim3)}.
When unspecified, uses
\code{image_data_format} value found in your Keras config file at
\verb{~/.keras/keras.json} (if exists) else 'channels_last'.
Defaults to 'channels_last'.}

\item{...}{standard layer arguments.}
}
\description{
Upsampling layer for 3D inputs
}
\details{
Repeats the 1st, 2nd and 3rd dimensions
of the data by \code{size[0]}, \code{size[1]} and \code{size[2]} respectively.

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> input_shape = (2, 1, 2, 1, 3)
>>> x = tf.constant(1, shape=input_shape)
>>> y = tf.keras.layers.UpSampling3D(size=2)(x)
>>> print(y.shape)
(2, 2, 4, 2, 3)
}\if{html}{\out{</div>}}

Input shape:
5D tensor with shape:
\itemize{
\item If \code{data_format} is \code{"channels_last"}:
\verb{(batch_size, dim1, dim2, dim3, channels)}
\item If \code{data_format} is \code{"channels_first"}:
\verb{(batch_size, channels, dim1, dim2, dim3)}
}

Output shape:
5D tensor with shape:
\itemize{
\item If \code{data_format} is \code{"channels_last"}:
\verb{(batch_size, upsampled_dim1, upsampled_dim2, upsampled_dim3, channels)}
\item If \code{data_format} is \code{"channels_first"}:
\verb{(batch_size, channels, upsampled_dim1, upsampled_dim2, upsampled_dim3)}
}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
