% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_wrapper}
\alias{layer_wrapper}
\title{Abstract wrapper base class}
\usage{
layer_wrapper(object, layer, ...)
}
\arguments{
\item{layer}{The layer to be wrapped.}

\item{...}{standard layer arguments.}
}
\description{
Abstract wrapper base class
}
\details{
Wrappers take another layer and augment it in various ways.
Do not use this class as a layer, it is only an abstract base class.
Two usable wrappers are the \code{TimeDistributed} and \code{Bidirectional} wrappers.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
