% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_zero_padding_1d}
\alias{layer_zero_padding_1d}
\title{Zero-padding layer for 1D input (e.g. temporal sequence)}
\usage{
layer_zero_padding_1d(object, padding = 1L, ...)
}
\arguments{
\item{padding}{Int, or list of int (length 2).
\itemize{
\item If int:
How many zeros to add at the beginning and end of
the padding dimension (axis 1).
\item If list of int (length 2):
How many zeros to add at the beginning and the end of
the padding dimension (\verb{(left_pad, right_pad)}).
}}

\item{...}{standard layer arguments.}
}
\description{
Zero-padding layer for 1D input (e.g. temporal sequence)
}
\details{
\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> input_shape = (2, 2, 3)
>>> x = np.arange(np.prod(input_shape)).reshape(input_shape)
>>> print(x)
[[[ 0  1  2]
  [ 3  4  5]]
 [[ 6  7  8]
  [ 9 10 11]]]
>>> y = tf.keras.layers.ZeroPadding1D(padding=2)(x)
>>> print(y)
tf.Tensor(
  [[[ 0  0  0]
    [ 0  0  0]
    [ 0  1  2]
    [ 3  4  5]
    [ 0  0  0]
    [ 0  0  0]]
   [[ 0  0  0]
    [ 0  0  0]
    [ 6  7  8]
    [ 9 10 11]
    [ 0  0  0]
    [ 0  0  0]]], shape=(2, 6, 3), dtype=int64)
}\if{html}{\out{</div>}}

Input shape:
3D tensor with shape \verb{(batch_size, axis_to_pad, features)}

Output shape:
3D tensor with shape \verb{(batch_size, padded_axis, features)}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
