% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-activations.R
\name{old_layer_activation}
\alias{old_layer_activation}
\title{Apply an activation function to an output.}
\usage{
old_layer_activation(
  object,
  activation,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{activation}{Activation function to use.
If you don't specify anything, no activation is applied
(ie. "linear" activation: \code{a(x) = x}).}

\item{input_shape}{Input shape (list of integers, does not include the
samples axis) which is required when using this layer as the first layer in
a model.}
}
\description{
Apply an activation function to an output.
}
\seealso{
Other core layers: 
\code{\link{old_layer_activity_regularization}()},
\code{\link{old_layer_attention}()},
\code{\link{old_layer_dense_features}()},
\code{\link{old_layer_dense}()},
\code{\link{old_layer_dropout}()},
\code{\link{old_layer_flatten}()},
\code{\link{old_layer_input}()},
\code{\link{old_layer_lambda}()},
\code{\link{old_layer_masking}()},
\code{\link{old_layer_permute}()},
\code{\link{old_layer_repeat_vector}()},
\code{\link{old_layer_reshape}()}

Other activation layers: 
\code{\link{old_layer_activation_elu}()},
\code{\link{old_layer_activation_leaky_relu}()},
\code{\link{old_layer_activation_parametric_relu}()},
\code{\link{old_layer_activation_relu}()},
\code{\link{old_layer_activation_selu}()},
\code{\link{old_layer_activation_softmax}()},
\code{\link{old_layer_activation_thresholded_relu}()}
}
\concept{activation layers}
\concept{core layers}
