% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-activations.R
\name{old_layer_activation_elu}
\alias{old_layer_activation_elu}
\title{Exponential Linear Unit.}
\usage{
old_layer_activation_elu(
  object,
  alpha = 1,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{alpha}{Scale for the negative factor.}
}
\description{
It follows: \code{f(x) =  alpha * (exp(x) - 1.0)} for \code{x < 0}, \code{f(x) = x} for \code{x >= 0}.
}
\seealso{
\href{https://arxiv.org/abs/1511.07289v1}{Fast and Accurate Deep Network Learning by Exponential Linear Units (ELUs)}.

Other activation layers: 
\code{\link{old_layer_activation_leaky_relu}()},
\code{\link{old_layer_activation_parametric_relu}()},
\code{\link{old_layer_activation_relu}()},
\code{\link{old_layer_activation_selu}()},
\code{\link{old_layer_activation_softmax}()},
\code{\link{old_layer_activation_thresholded_relu}()},
\code{\link{old_layer_activation}()}
}
\concept{activation layers}
