% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-activations.R
\name{old_layer_activation_leaky_relu}
\alias{old_layer_activation_leaky_relu}
\title{Leaky version of a Rectified Linear Unit.}
\usage{
old_layer_activation_leaky_relu(
  object,
  alpha = 0.3,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{alpha}{float >= 0. Negative slope coefficient.}
}
\description{
Allows a small gradient when the unit is not active: \code{f(x) = alpha * x} for
\code{x < 0}, \code{f(x) = x} for \code{x >= 0}.
}
\seealso{
\href{https://ai.stanford.edu/~amaas/papers/relu_hybrid_icml2013_final.pdf}{Rectifier Nonlinearities Improve Neural Network Acoustic Models}.

Other activation layers: 
\code{\link{old_layer_activation_elu}()},
\code{\link{old_layer_activation_parametric_relu}()},
\code{\link{old_layer_activation_relu}()},
\code{\link{old_layer_activation_selu}()},
\code{\link{old_layer_activation_softmax}()},
\code{\link{old_layer_activation_thresholded_relu}()},
\code{\link{old_layer_activation}()}
}
\concept{activation layers}
