% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-activations.R
\name{old_layer_activation_parametric_relu}
\alias{old_layer_activation_parametric_relu}
\title{Parametric Rectified Linear Unit.}
\usage{
old_layer_activation_parametric_relu(
  object,
  alpha_initializer = "zeros",
  alpha_regularizer = NULL,
  alpha_constraint = NULL,
  shared_axes = NULL,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{alpha_initializer}{Initializer function for the weights.}

\item{alpha_regularizer}{Regularizer for the weights.}

\item{alpha_constraint}{Constraint for the weights.}

\item{shared_axes}{The axes along which to share learnable parameters for the
activation function. For example, if the incoming feature maps are from a
2D convolution with output shape (batch, height, width, channels), and you
wish to share parameters across space so that each filter only has one set
of parameters, set shared_axes=c(1, 2).}
}
\description{
It follows: \verb{f(x) = alpha * x`` for }x < 0\verb{, }f(x) = x\code{for}x >= 0`, where
alpha is a learned array with the same shape as x.
}
\seealso{
\href{https://arxiv.org/abs/1502.01852}{Delving Deep into Rectifiers: Surpassing Human-Level Performance on ImageNet Classification}.

Other activation layers: 
\code{\link{old_layer_activation_elu}()},
\code{\link{old_layer_activation_leaky_relu}()},
\code{\link{old_layer_activation_relu}()},
\code{\link{old_layer_activation_selu}()},
\code{\link{old_layer_activation_softmax}()},
\code{\link{old_layer_activation_thresholded_relu}()},
\code{\link{old_layer_activation}()}
}
\concept{activation layers}
