% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-activations.R
\name{old_layer_activation_relu}
\alias{old_layer_activation_relu}
\title{Rectified Linear Unit activation function}
\usage{
old_layer_activation_relu(
  object,
  max_value = NULL,
  negative_slope = 0,
  threshold = 0,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{max_value}{loat, the maximum output value.}

\item{negative_slope}{float >= 0 Negative slope coefficient.}

\item{threshold}{float. Threshold value for thresholded activation.}
}
\description{
Rectified Linear Unit activation function
}
\seealso{
Other activation layers: 
\code{\link{old_layer_activation_elu}()},
\code{\link{old_layer_activation_leaky_relu}()},
\code{\link{old_layer_activation_parametric_relu}()},
\code{\link{old_layer_activation_selu}()},
\code{\link{old_layer_activation_softmax}()},
\code{\link{old_layer_activation_thresholded_relu}()},
\code{\link{old_layer_activation}()}
}
\concept{activation layers}
