% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-activations.R
\name{old_layer_activation_selu}
\alias{old_layer_activation_selu}
\title{Scaled Exponential Linear Unit.}
\usage{
old_layer_activation_selu(
  object,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\description{
SELU is equal to: \code{scale * elu(x, alpha)}, where alpha and scale
are pre-defined constants.
}
\details{
The values of \code{alpha} and \code{scale} are
chosen so that the mean and variance of the inputs are preserved
between two consecutive layers as long as the weights are initialized
correctly (see initializer_lecun_normal) and the number of inputs
is "large enough" (see article for more information).

Note:
\itemize{
\item To be used together with the initialization "lecun_normal".
\item To be used together with the dropout variant "AlphaDropout".
}
}
\seealso{
\href{https://arxiv.org/abs/1706.02515}{Self-Normalizing Neural Networks}, \code{\link{initializer_lecun_normal}}, \code{\link{layer_alpha_dropout}}

Other activation layers: 
\code{\link{old_layer_activation_elu}()},
\code{\link{old_layer_activation_leaky_relu}()},
\code{\link{old_layer_activation_parametric_relu}()},
\code{\link{old_layer_activation_relu}()},
\code{\link{old_layer_activation_softmax}()},
\code{\link{old_layer_activation_thresholded_relu}()},
\code{\link{old_layer_activation}()}
}
\concept{activation layers}
