% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-activations.R
\name{old_layer_activation_softmax}
\alias{old_layer_activation_softmax}
\title{Softmax activation function.}
\usage{
old_layer_activation_softmax(
  object,
  axis = -1,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{axis}{Integer, axis along which the softmax normalization is applied.}
}
\description{
It follows: \code{f(x) =  alpha * (exp(x) - 1.0)} for \code{x < 0}, \code{f(x) = x} for \code{x >= 0}.
}
\seealso{
Other activation layers: 
\code{\link{old_layer_activation_elu}()},
\code{\link{old_layer_activation_leaky_relu}()},
\code{\link{old_layer_activation_parametric_relu}()},
\code{\link{old_layer_activation_relu}()},
\code{\link{old_layer_activation_selu}()},
\code{\link{old_layer_activation_thresholded_relu}()},
\code{\link{old_layer_activation}()}
}
\concept{activation layers}
