% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-activations.R
\name{old_layer_activation_thresholded_relu}
\alias{old_layer_activation_thresholded_relu}
\title{Thresholded Rectified Linear Unit.}
\usage{
old_layer_activation_thresholded_relu(
  object,
  theta = 1,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{theta}{float >= 0. Threshold location of activation.}
}
\description{
It follows: \code{f(x) = x} for \code{x > theta}, \code{f(x) = 0} otherwise.
}
\seealso{
\href{https://arxiv.org/abs/1402.3337}{Zero-bias autoencoders and the benefits of co-adapting features}.

Other activation layers: 
\code{\link{old_layer_activation_elu}()},
\code{\link{old_layer_activation_leaky_relu}()},
\code{\link{old_layer_activation_parametric_relu}()},
\code{\link{old_layer_activation_relu}()},
\code{\link{old_layer_activation_selu}()},
\code{\link{old_layer_activation_softmax}()},
\code{\link{old_layer_activation}()}
}
\concept{activation layers}
