% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-pooling.R
\name{old_layer_average_pooling_1d}
\alias{old_layer_average_pooling_1d}
\title{Average pooling for temporal data.}
\usage{
old_layer_average_pooling_1d(
  object,
  pool_size = 2L,
  strides = NULL,
  padding = "valid",
  data_format = "channels_last",
  batch_size = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{pool_size}{Integer, size of the average pooling windows.}

\item{strides}{Integer, or NULL. Factor by which to downscale. E.g. 2 will
halve the input. If NULL, it will default to \code{pool_size}.}

\item{padding}{One of \code{"valid"} or \code{"same"} (case-insensitive).}

\item{data_format}{One of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.}
}
\description{
Average pooling for temporal data.
}
\section{Input shape}{
 3D tensor with shape: \verb{(batch_size, steps, features)}.
}

\section{Output shape}{
 3D tensor with shape: \verb{(batch_size, downsampled_steps, features)}.
}

\seealso{
Other pooling layers: 
\code{\link{old_layer_average_pooling_2d}()},
\code{\link{old_layer_average_pooling_3d}()},
\code{\link{old_layer_global_average_pooling_1d}()},
\code{\link{old_layer_global_average_pooling_2d}()},
\code{\link{old_layer_global_average_pooling_3d}()},
\code{\link{old_layer_global_max_pooling_1d}()},
\code{\link{old_layer_global_max_pooling_2d}()},
\code{\link{old_layer_global_max_pooling_3d}()},
\code{\link{old_layer_max_pooling_1d}()},
\code{\link{old_layer_max_pooling_2d}()},
\code{\link{old_layer_max_pooling_3d}()}
}
\concept{pooling layers}
