% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-pooling.R
\name{old_layer_average_pooling_3d}
\alias{old_layer_average_pooling_3d}
\title{Average pooling operation for 3D data (spatial or spatio-temporal).}
\usage{
old_layer_average_pooling_3d(
  object,
  pool_size = c(2L, 2L, 2L),
  strides = NULL,
  padding = "valid",
  data_format = NULL,
  batch_size = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{pool_size}{list of 3 integers, factors by which to downscale (dim1,
dim2, dim3). (2, 2, 2) will halve the size of the 3D input in each
dimension.}

\item{strides}{list of 3 integers, or NULL. Strides values.}

\item{padding}{One of \code{"valid"} or \code{"same"} (case-insensitive).}

\item{data_format}{A string, one of \code{channels_last} (default) or
\code{channels_first}. The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape \verb{(batch, spatial_dim1, spatial_dim2, spatial_dim3, channels)} while \code{channels_first} corresponds
to inputs with shape \verb{(batch, channels, spatial_dim1, spatial_dim2, spatial_dim3)}. It defaults to the \code{image_data_format} value found in your
Keras config file at \verb{~/.keras/keras.json}. If you never set it, then it
will be "channels_last".}
}
\description{
Average pooling operation for 3D data (spatial or spatio-temporal).
}
\section{Input shape}{

\itemize{
\item If \code{data_format='channels_last'}: 5D tensor with shape: \verb{(batch_size, spatial_dim1, spatial_dim2, spatial_dim3, channels)}
\item If \code{data_format='channels_first'}: 5D tensor with shape: \verb{(batch_size, channels, spatial_dim1, spatial_dim2, spatial_dim3)}
}
}

\section{Output shape}{

\itemize{
\item If \code{data_format='channels_last'}: 5D tensor with shape: \verb{(batch_size, pooled_dim1, pooled_dim2, pooled_dim3, channels)}
\item If \code{data_format='channels_first'}: 5D tensor with shape: \verb{(batch_size, channels, pooled_dim1, pooled_dim2, pooled_dim3)}
}
}

\seealso{
Other pooling layers: 
\code{\link{old_layer_average_pooling_1d}()},
\code{\link{old_layer_average_pooling_2d}()},
\code{\link{old_layer_global_average_pooling_1d}()},
\code{\link{old_layer_global_average_pooling_2d}()},
\code{\link{old_layer_global_average_pooling_3d}()},
\code{\link{old_layer_global_max_pooling_1d}()},
\code{\link{old_layer_global_max_pooling_2d}()},
\code{\link{old_layer_global_max_pooling_3d}()},
\code{\link{old_layer_max_pooling_1d}()},
\code{\link{old_layer_max_pooling_2d}()},
\code{\link{old_layer_max_pooling_3d}()}
}
\concept{pooling layers}
