% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-preprocessing.R
\name{old_layer_center_crop}
\alias{old_layer_center_crop}
\title{Crop the central portion of the images to target height and width}
\usage{
old_layer_center_crop(object, height, width, ...)
}
\arguments{
\item{height}{Integer, the height of the output shape.}

\item{width}{Integer, the width of the output shape.}

\item{...}{standard layer arguments.}
}
\description{
Crop the central portion of the images to target height and width
}
\details{
Input shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., height, width, channels)}, in \code{"channels_last"} format.

Output shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., target_height, target_width, channels)}.

If the input height/width is even and the target height/width is odd (or
inversely), the input image is left-padded by 1 pixel.
}
\seealso{
\itemize{
\item \url{https://www.tensorflow.org/api_docs/python/tf/keras/layers/CenterCrop}
\item \url{https://keras.io/api/layers/preprocessing_layers/image_preprocessing/center_crop}
}

Other image preprocessing layers: 
\code{\link{old_layer_rescaling}()},
\code{\link{old_layer_resizing}()}

Other preprocessing layers: 
\code{\link{old_layer_category_encoding}()},
\code{\link{old_layer_discretization}()},
\code{\link{old_layer_hashing}()},
\code{\link{old_layer_integer_lookup}()},
\code{\link{old_layer_normalization}()},
\code{\link{old_layer_random_brightness}()},
\code{\link{old_layer_random_contrast}()},
\code{\link{old_layer_random_crop}()},
\code{\link{old_layer_random_flip}()},
\code{\link{old_layer_random_height}()},
\code{\link{old_layer_random_rotation}()},
\code{\link{old_layer_random_translation}()},
\code{\link{old_layer_random_width}()},
\code{\link{old_layer_random_zoom}()},
\code{\link{old_layer_rescaling}()},
\code{\link{old_layer_resizing}()},
\code{\link{old_layer_string_lookup}()},
\code{\link{old_layer_text_vectorization}()}
}
\concept{image preprocessing layers}
\concept{preprocessing layers}
