% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-merge.R
\name{old_layer_concatenate}
\alias{old_layer_concatenate}
\title{Layer that concatenates a list of inputs.}
\usage{
old_layer_concatenate(inputs, ..., axis = -1)
}
\arguments{
\item{inputs}{A input tensor, or list of input tensors. Can be missing.}

\item{...}{Unnamed args are treated as additional \code{inputs}. Named arguments are passed on as standard layer arguments.}

\item{axis}{Concatenation axis.}
}
\value{
A tensor, the concatenation of the inputs alongside axis \code{axis}. If
\code{inputs} is missing, a keras layer instance is returned.
}
\description{
It takes as input a list of tensors, all of the same shape expect for the
concatenation axis, and returns a single tensor, the concatenation of all
inputs.
}
\seealso{
\itemize{
\item \url{https://www.tensorflow.org/api_docs/python/tf/keras/layers/concatenate}
\item \url{https://www.tensorflow.org/api_docs/python/tf/keras/layers/Concatenate}
\item \url{https://keras.io/api/layers/merging_layers/concatenate}
}

Other merge layers: 
\code{\link{old_layer_average}()},
\code{\link{old_layer_dot}()},
\code{\link{old_layer_maximum}()},
\code{\link{old_layer_minimum}()},
\code{\link{old_layer_multiply}()},
\code{\link{old_layer_subtract}()}
}
\concept{merge layers}
