% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-features.R
\name{old_layer_dense_features}
\alias{old_layer_dense_features}
\title{Constructs a DenseFeatures.}
\usage{
old_layer_dense_features(
  object,
  feature_columns,
  name = NULL,
  trainable = NULL,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  weights = NULL
)
}
\arguments{
\item{feature_columns}{An iterable containing the FeatureColumns to use as
inputs to your model. All items should be instances of classes derived from
\code{DenseColumn} such as \code{numeric_column}, \code{embedding_column}, \code{bucketized_column},
\code{indicator_column}. If you have categorical features, you can wrap them with an
\code{embedding_column} or \code{indicator_column}. See \code{tfestimators::feature_columns()}.}
}
\description{
A layer that produces a dense Tensor based on given feature_columns.
}
\seealso{
Other core layers: 
\code{\link{old_layer_activation}()},
\code{\link{old_layer_activity_regularization}()},
\code{\link{old_layer_attention}()},
\code{\link{old_layer_dense}()},
\code{\link{old_layer_dropout}()},
\code{\link{old_layer_flatten}()},
\code{\link{old_layer_input}()},
\code{\link{old_layer_lambda}()},
\code{\link{old_layer_masking}()},
\code{\link{old_layer_permute}()},
\code{\link{old_layer_repeat_vector}()},
\code{\link{old_layer_reshape}()}
}
\concept{core layers}
