% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-merge.R
\name{old_layer_dot}
\alias{old_layer_dot}
\title{Layer that computes a dot product between samples in two tensors.}
\usage{
old_layer_dot(inputs, ..., axes, normalize = FALSE)
}
\arguments{
\item{inputs}{A input tensor, or list of input tensors. Can be missing.}

\item{...}{Unnamed args are treated as additional \code{inputs}. Named arguments are passed on as standard layer arguments.}

\item{axes}{Integer or list of integers, axis or axes along which to take the
dot product.}

\item{normalize}{Whether to L2-normalize samples along the dot product axis
before taking the dot product. If set to TRUE, then the output of the dot
product is the cosine proximity between the two samples.}
}
\value{
If \code{inputs} is supplied: A tensor, the dot product of the samples
from the inputs. If \code{inputs} is missing, a keras layer instance is
returned.
}
\description{
Layer that computes a dot product between samples in two tensors.
}
\seealso{
\itemize{
\item \url{https://www.tensorflow.org/api_docs/python/tf/keras/layers/dot}
\item \url{https://keras.io/api/layers/merging_layers/dot/}
}

Other merge layers: 
\code{\link{old_layer_average}()},
\code{\link{old_layer_concatenate}()},
\code{\link{old_layer_maximum}()},
\code{\link{old_layer_minimum}()},
\code{\link{old_layer_multiply}()},
\code{\link{old_layer_subtract}()}
}
\concept{merge layers}
