% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-core.R
\name{old_layer_flatten}
\alias{old_layer_flatten}
\title{Flattens an input}
\usage{
old_layer_flatten(
  object,
  data_format = NULL,
  input_shape = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{data_format}{A string. one of \code{channels_last} (default) or
\code{channels_first}. The ordering of the dimensions in the inputs. The purpose
of this argument is to preserve weight ordering when switching a model from
one data format to another. \code{channels_last} corresponds to inputs with
shape \verb{(batch, ..., channels)} while \code{channels_first} corresponds to inputs
with shape \verb{(batch, channels, ...)}. It defaults to the \code{image_data_format}
value found in your Keras config file at \verb{~/.keras/keras.json}. If you
never set it, then it will be "channels_last".}
}
\description{
Flatten a given input, does not affect the batch size.
}
\seealso{
Other core layers: 
\code{\link{old_layer_activation}()},
\code{\link{old_layer_activity_regularization}()},
\code{\link{old_layer_attention}()},
\code{\link{old_layer_dense_features}()},
\code{\link{old_layer_dense}()},
\code{\link{old_layer_dropout}()},
\code{\link{old_layer_input}()},
\code{\link{old_layer_lambda}()},
\code{\link{old_layer_masking}()},
\code{\link{old_layer_permute}()},
\code{\link{old_layer_repeat_vector}()},
\code{\link{old_layer_reshape}()}
}
\concept{core layers}
