% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-pooling.R
\name{old_layer_global_max_pooling_1d}
\alias{old_layer_global_max_pooling_1d}
\title{Global max pooling operation for temporal data.}
\usage{
old_layer_global_max_pooling_1d(
  object,
  data_format = "channels_last",
  keepdims = FALSE,
  ...
)
}
\arguments{
\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch, steps, features)} while \code{channels_first}
corresponds to inputs with shape
\verb{(batch, features, steps)}.}

\item{keepdims}{A boolean, whether to keep the spatial dimensions or not. If
\code{keepdims} is \code{FALSE} (default), the rank of the tensor is reduced for
spatial dimensions. If \code{keepdims} is \code{TRUE}, the spatial dimensions are
retained with length 1. The behavior is the same as for \code{tf.reduce_mean} or
\code{np.mean}.}

\item{...}{standard layer arguments.}
}
\description{
Global max pooling operation for temporal data.
}
\section{Input shape}{

3D tensor with shape: \verb{(batch_size, steps, features)}.
}

\section{Output shape}{

2D tensor with shape: \verb{(batch_size, channels)}
}

\seealso{
Other pooling layers: 
\code{\link{old_layer_average_pooling_1d}()},
\code{\link{old_layer_average_pooling_2d}()},
\code{\link{old_layer_average_pooling_3d}()},
\code{\link{old_layer_global_average_pooling_1d}()},
\code{\link{old_layer_global_average_pooling_2d}()},
\code{\link{old_layer_global_average_pooling_3d}()},
\code{\link{old_layer_global_max_pooling_2d}()},
\code{\link{old_layer_global_max_pooling_3d}()},
\code{\link{old_layer_max_pooling_1d}()},
\code{\link{old_layer_max_pooling_2d}()},
\code{\link{old_layer_max_pooling_3d}()}
}
\concept{pooling layers}
