% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-pooling.R
\name{old_layer_global_max_pooling_2d}
\alias{old_layer_global_max_pooling_2d}
\title{Global max pooling operation for spatial data.}
\usage{
old_layer_global_max_pooling_2d(
  object,
  data_format = NULL,
  keepdims = FALSE,
  ...
)
}
\arguments{
\item{data_format}{A string, one of \code{channels_last} (default) or
\code{channels_first}.  The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape \verb{(batch_size, height, width, channels)} while \code{channels_first} corresponds to inputs with
shape \verb{(batch_size, channels, height, width)}. If left unspecified, it
uses the \code{image_data_format} value found in your Keras config file at
\verb{~/.keras/keras.json} (if exists) else 'channels_last'.
Note that the \code{channels_first} format is currently not
supported by TensorFlow on CPU. Defaults to 'channels_last'.}

\item{keepdims}{A boolean, whether to keep the temporal dimension or not.
If \code{keepdims} is \code{FALSE} (default), the rank of the tensor is reduced
for spatial dimensions.
If \code{keepdims} is \code{TRUE}, the temporal dimension are retained with
length 1.
The behavior is the same as for \code{tf.reduce_max} or \code{np.max}.}

\item{...}{standard layer arguments.}
}
\description{
Global max pooling operation for spatial data.
}
\section{Input shape}{

\itemize{
\item If \code{data_format='channels_last'}: 4D tensor with shape: \verb{(batch_size, rows, cols, channels)}
\item If \code{data_format='channels_first'}: 4D tensor with shape: \verb{(batch_size, channels, rows, cols)}
}
}

\section{Output shape}{
 2D tensor with shape: \verb{(batch_size, channels)}
}

\seealso{
Other pooling layers: 
\code{\link{old_layer_average_pooling_1d}()},
\code{\link{old_layer_average_pooling_2d}()},
\code{\link{old_layer_average_pooling_3d}()},
\code{\link{old_layer_global_average_pooling_1d}()},
\code{\link{old_layer_global_average_pooling_2d}()},
\code{\link{old_layer_global_average_pooling_3d}()},
\code{\link{old_layer_global_max_pooling_1d}()},
\code{\link{old_layer_global_max_pooling_3d}()},
\code{\link{old_layer_max_pooling_1d}()},
\code{\link{old_layer_max_pooling_2d}()},
\code{\link{old_layer_max_pooling_3d}()}
}
\concept{pooling layers}
