% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-core.R
\name{old_layer_lambda}
\alias{old_layer_lambda}
\title{Wraps arbitrary expression as a layer}
\usage{
old_layer_lambda(
  object,
  f,
  output_shape = NULL,
  mask = NULL,
  arguments = NULL,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{f}{The function to be evaluated. Takes input tensor as first
argument.}

\item{output_shape}{Expected output shape from the function (not required
when using TensorFlow back-end).}

\item{mask}{mask}

\item{arguments}{optional named list of keyword arguments to be passed to the
function.}
}
\description{
Wraps arbitrary expression as a layer
}
\section{Input shape}{
 Arbitrary. Use the keyword argument input_shape (list
of integers, does not include the samples axis) when using this layer as
the first layer in a model.
}

\section{Output shape}{
 Arbitrary (based on tensor returned from the function)
}

\seealso{
Other core layers: 
\code{\link{old_layer_activation}()},
\code{\link{old_layer_activity_regularization}()},
\code{\link{old_layer_attention}()},
\code{\link{old_layer_dense_features}()},
\code{\link{old_layer_dense}()},
\code{\link{old_layer_dropout}()},
\code{\link{old_layer_flatten}()},
\code{\link{old_layer_input}()},
\code{\link{old_layer_masking}()},
\code{\link{old_layer_permute}()},
\code{\link{old_layer_repeat_vector}()},
\code{\link{old_layer_reshape}()}
}
\concept{core layers}
