% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-core.R
\name{old_layer_masking}
\alias{old_layer_masking}
\title{Masks a sequence by using a mask value to skip timesteps.}
\usage{
old_layer_masking(
  object,
  mask_value = 0,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{mask_value}{float, mask value}
}
\description{
For each timestep in the input tensor (dimension #1 in the tensor), if all
values in the input tensor at that timestep are equal to \code{mask_value}, then
the timestep will be masked (skipped) in all downstream layers (as long as
they support masking). If any downstream layer does not support masking yet
receives such an input mask, an exception will be raised.
}
\seealso{
Other core layers: 
\code{\link{old_layer_activation}()},
\code{\link{old_layer_activity_regularization}()},
\code{\link{old_layer_attention}()},
\code{\link{old_layer_dense_features}()},
\code{\link{old_layer_dense}()},
\code{\link{old_layer_dropout}()},
\code{\link{old_layer_flatten}()},
\code{\link{old_layer_input}()},
\code{\link{old_layer_lambda}()},
\code{\link{old_layer_permute}()},
\code{\link{old_layer_repeat_vector}()},
\code{\link{old_layer_reshape}()}
}
\concept{core layers}
