% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-core.R
\name{old_layer_permute}
\alias{old_layer_permute}
\title{Permute the dimensions of an input according to a given pattern}
\usage{
old_layer_permute(
  object,
  dims,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{dims}{List of integers. Permutation pattern, does not include the
samples dimension. Indexing starts at 1. For instance, \verb{(2, 1)} permutes
the first and second dimension of the input.}
}
\description{
Permute the dimensions of an input according to a given pattern
}
\note{
Useful for e.g. connecting RNNs and convnets together.
}
\section{Input and Output Shapes}{


Input shape: Arbitrary

Output shape: Same as the input shape, but with the dimensions re-ordered
according to the specified pattern.
}

\seealso{
Other core layers: 
\code{\link{old_layer_activation}()},
\code{\link{old_layer_activity_regularization}()},
\code{\link{old_layer_attention}()},
\code{\link{old_layer_dense_features}()},
\code{\link{old_layer_dense}()},
\code{\link{old_layer_dropout}()},
\code{\link{old_layer_flatten}()},
\code{\link{old_layer_input}()},
\code{\link{old_layer_lambda}()},
\code{\link{old_layer_masking}()},
\code{\link{old_layer_repeat_vector}()},
\code{\link{old_layer_reshape}()}
}
\concept{core layers}
