% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-preprocessing.R
\name{old_layer_random_flip}
\alias{old_layer_random_flip}
\title{Randomly flip each image horizontally and vertically}
\usage{
old_layer_random_flip(
  object,
  mode = "horizontal_and_vertical",
  seed = NULL,
  ...
)
}
\arguments{
\item{mode}{String indicating which flip mode to use. Can be \code{"horizontal"},
\code{"vertical"}, or \code{"horizontal_and_vertical"}. Defaults to
\code{"horizontal_and_vertical"}. \code{"horizontal"} is a left-right flip and
\code{"vertical"} is a top-bottom flip.}

\item{seed}{Integer. Used to create a random seed.}

\item{...}{standard layer arguments.}
}
\description{
Randomly flip each image horizontally and vertically
}
\details{
This layer will flip the images based on the \code{mode} attribute.
During inference time, the output will be identical to input. Call the layer
with \code{training = TRUE} to flip the input.

Input shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., height, width, channels)}, in \code{"channels_last"} format.

Output shape:
3D (unbatched) or 4D (batched) tensor with shape:
\verb{(..., height, width, channels)}, in \code{"channels_last"} format.
}
\seealso{
\itemize{
\item \url{https://www.tensorflow.org/api_docs/python/tf/keras/layers/RandomFlip}
\item \url{https://keras.io/api/layers/preprocessing_layers/image_augmentation/random_flip}
}

Other image augmentation layers: 
\code{\link{old_layer_random_brightness}()},
\code{\link{old_layer_random_contrast}()},
\code{\link{old_layer_random_crop}()},
\code{\link{old_layer_random_height}()},
\code{\link{old_layer_random_rotation}()},
\code{\link{old_layer_random_translation}()},
\code{\link{old_layer_random_width}()},
\code{\link{old_layer_random_zoom}()}

Other preprocessing layers: 
\code{\link{old_layer_category_encoding}()},
\code{\link{old_layer_center_crop}()},
\code{\link{old_layer_discretization}()},
\code{\link{old_layer_hashing}()},
\code{\link{old_layer_integer_lookup}()},
\code{\link{old_layer_normalization}()},
\code{\link{old_layer_random_brightness}()},
\code{\link{old_layer_random_contrast}()},
\code{\link{old_layer_random_crop}()},
\code{\link{old_layer_random_height}()},
\code{\link{old_layer_random_rotation}()},
\code{\link{old_layer_random_translation}()},
\code{\link{old_layer_random_width}()},
\code{\link{old_layer_random_zoom}()},
\code{\link{old_layer_rescaling}()},
\code{\link{old_layer_resizing}()},
\code{\link{old_layer_string_lookup}()},
\code{\link{old_layer_text_vectorization}()}
}
\concept{image augmentation layers}
\concept{preprocessing layers}
