% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-core.R
\name{old_layer_repeat_vector}
\alias{old_layer_repeat_vector}
\title{Repeats the input n times.}
\usage{
old_layer_repeat_vector(
  object,
  n,
  batch_size = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{n}{integer, repetition factor.}
}
\description{
Repeats the input n times.
}
\section{Input shape}{
 2D tensor of shape \verb{(num_samples, features)}.
}

\section{Output shape}{
 3D tensor of shape \verb{(num_samples, n, features)}.
}

\seealso{
Other core layers: 
\code{\link{old_layer_activation}()},
\code{\link{old_layer_activity_regularization}()},
\code{\link{old_layer_attention}()},
\code{\link{old_layer_dense_features}()},
\code{\link{old_layer_dense}()},
\code{\link{old_layer_dropout}()},
\code{\link{old_layer_flatten}()},
\code{\link{old_layer_input}()},
\code{\link{old_layer_lambda}()},
\code{\link{old_layer_masking}()},
\code{\link{old_layer_permute}()},
\code{\link{old_layer_reshape}()}
}
\concept{core layers}
