% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-preprocessing.R
\name{old_layer_rescaling}
\alias{old_layer_rescaling}
\title{Multiply inputs by \code{scale} and adds \code{offset}}
\usage{
old_layer_rescaling(object, scale, offset = 0, ...)
}
\arguments{
\item{scale}{Float, the scale to apply to the inputs.}

\item{offset}{Float, the offset to apply to the inputs.}

\item{...}{standard layer arguments.}
}
\description{
Multiply inputs by \code{scale} and adds \code{offset}
}
\details{
For instance:
\enumerate{
\item To rescale an input in the \verb{[0, 255]} range
to be in the \verb{[0, 1]} range, you would pass \code{scale=1./255}.
\item To rescale an input in the \verb{[0, 255]} range to be in the \verb{[-1, 1]} range,
you would pass \verb{scale = 1/127.5, offset = -1}.
}

The rescaling is applied both during training and inference.

Input shape:
Arbitrary.

Output shape:
Same as input.
}
\seealso{
\itemize{
\item \url{https://www.tensorflow.org/api_docs/python/tf/keras/layers/Rescaling}
\item \url{https://keras.io/api/layers/preprocessing_layers/image_preprocessing/rescaling}
}

Other image preprocessing layers: 
\code{\link{old_layer_center_crop}()},
\code{\link{old_layer_resizing}()}

Other preprocessing layers: 
\code{\link{old_layer_category_encoding}()},
\code{\link{old_layer_center_crop}()},
\code{\link{old_layer_discretization}()},
\code{\link{old_layer_hashing}()},
\code{\link{old_layer_integer_lookup}()},
\code{\link{old_layer_normalization}()},
\code{\link{old_layer_random_brightness}()},
\code{\link{old_layer_random_contrast}()},
\code{\link{old_layer_random_crop}()},
\code{\link{old_layer_random_flip}()},
\code{\link{old_layer_random_height}()},
\code{\link{old_layer_random_rotation}()},
\code{\link{old_layer_random_translation}()},
\code{\link{old_layer_random_width}()},
\code{\link{old_layer_random_zoom}()},
\code{\link{old_layer_resizing}()},
\code{\link{old_layer_string_lookup}()},
\code{\link{old_layer_text_vectorization}()}
}
\concept{image preprocessing layers}
\concept{preprocessing layers}
