% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-core.R
\name{old_layer_reshape}
\alias{old_layer_reshape}
\title{Reshapes an output to a certain shape.}
\usage{
old_layer_reshape(
  object,
  target_shape,
  input_shape = NULL,
  batch_input_shape = NULL,
  batch_size = NULL,
  dtype = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{target_shape}{List of integers, does not include the samples dimension
(batch size).}
}
\description{
Reshapes an output to a certain shape.
}
\section{Input and Output Shapes}{


Input shape: Arbitrary, although all dimensions in the input shaped must be
fixed.

Output shape: \verb{(batch_size,) + target_shape}.
}

\seealso{
Other core layers: 
\code{\link{old_layer_activation}()},
\code{\link{old_layer_activity_regularization}()},
\code{\link{old_layer_attention}()},
\code{\link{old_layer_dense_features}()},
\code{\link{old_layer_dense}()},
\code{\link{old_layer_dropout}()},
\code{\link{old_layer_flatten}()},
\code{\link{old_layer_input}()},
\code{\link{old_layer_lambda}()},
\code{\link{old_layer_masking}()},
\code{\link{old_layer_permute}()},
\code{\link{old_layer_repeat_vector}()}
}
\concept{core layers}
