% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-dropout.R
\name{old_layer_spatial_dropout_1d}
\alias{old_layer_spatial_dropout_1d}
\title{Spatial 1D version of Dropout.}
\usage{
old_layer_spatial_dropout_1d(
  object,
  rate,
  batch_size = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{rate}{Float between 0 and 1. Fraction of the input units to drop.}
}
\description{
This version performs the same function as Dropout, however it drops entire
1D feature maps instead of individual elements. If adjacent frames within
feature maps are strongly correlated (as is normally the case in early
convolution layers) then regular dropout will not regularize the activations
and will otherwise just result in an effective learning rate decrease. In
this case, \code{layer_spatial_dropout_1d} will help promote independence between
feature maps and should be used instead.
}
\section{Input shape}{
 3D tensor with shape: \verb{(samples, timesteps, channels)}
}

\section{Output shape}{
 Same as input
}

\section{References}{
 - \href{https://arxiv.org/abs/1411.4280}{Efficient Object Localization Using Convolutional Networks}
}

\seealso{
Other dropout layers: 
\code{\link{old_layer_dropout}()},
\code{\link{old_layer_spatial_dropout_2d}()},
\code{\link{old_layer_spatial_dropout_3d}()}
}
\concept{dropout layers}
