% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-recurrent-cells.R
\name{old_layer_stacked_rnn_cells}
\alias{old_layer_stacked_rnn_cells}
\title{Wrapper allowing a stack of RNN cells to behave as a single cell}
\usage{
old_layer_stacked_rnn_cells(cells, ...)
}
\arguments{
\item{cells}{List of RNN cell instances.}

\item{...}{standard layer arguments.}
}
\description{
Used to implement efficient stacked RNNs.
}
\seealso{
\itemize{
\item \url{https://www.tensorflow.org/api_docs/python/tf/keras/layers/StackedRNNCells}
}

Other RNN cell layers: 
\code{\link{old_layer_gru_cell}()},
\code{\link{old_layer_lstm_cell}()},
\code{\link{old_layer_simple_rnn_cell}()}
}
\concept{RNN cell layers}
