% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-convolutional.R
\name{old_layer_zero_padding_1d}
\alias{old_layer_zero_padding_1d}
\title{Zero-padding layer for 1D input (e.g. temporal sequence).}
\usage{
old_layer_zero_padding_1d(
  object,
  padding = 1L,
  batch_size = NULL,
  name = NULL,
  trainable = NULL,
  weights = NULL
)
}
\arguments{
\item{padding}{int, or list of int (length 2)
\itemize{
\item If int: How many zeros to add at the beginning and end of the padding dimension (axis 1).
\item If list of int (length 2): How many zeros to add at the beginning and at the end of the padding dimension (\verb{(left_pad, right_pad)}).
}}
}
\description{
Zero-padding layer for 1D input (e.g. temporal sequence).
}
\section{Input shape}{

3D tensor with shape \verb{(batch, axis_to_pad, features)}
}

\section{Output shape}{

3D tensor with shape \verb{(batch, padded_axis, features)}
}

\seealso{
Other convolutional layers: 
\code{\link{old_layer_conv_1d_transpose}()},
\code{\link{old_layer_conv_1d}()},
\code{\link{old_layer_conv_2d_transpose}()},
\code{\link{old_layer_conv_2d}()},
\code{\link{old_layer_conv_3d_transpose}()},
\code{\link{old_layer_conv_3d}()},
\code{\link{old_layer_conv_lstm_2d}()},
\code{\link{old_layer_cropping_1d}()},
\code{\link{old_layer_cropping_2d}()},
\code{\link{old_layer_cropping_3d}()},
\code{\link{old_layer_depthwise_conv_1d}()},
\code{\link{old_layer_depthwise_conv_2d}()},
\code{\link{old_layer_separable_conv_1d}()},
\code{\link{old_layer_separable_conv_2d}()},
\code{\link{old_layer_upsampling_1d}()},
\code{\link{old_layer_upsampling_2d}()},
\code{\link{old_layer_upsampling_3d}()},
\code{\link{old_layer_zero_padding_2d}()},
\code{\link{old_layer_zero_padding_3d}()}
}
\concept{convolutional layers}
