% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_activation}
\alias{layer_activation}
\title{Applies an activation function to an output}
\usage{
layer_activation(object, activation, ...)
}
\arguments{
\item{activation}{Activation function, such as \code{tf.nn.relu}, or string name of
built-in activation function, such as "relu".}

\item{...}{standard layer arguments.}
}
\description{
Applies an activation function to an output
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
