% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_add}
\alias{layer_add}
\title{Layer that adds a list of inputs}
\usage{
layer_add(inputs, ...)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Layer that adds a list of inputs
}
\details{
It takes as input a list of tensors,
all of the same shape, and returns
a single tensor (also of the same shape).

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> input_shape = (2, 3, 4)
>>> x1 = tf.random.normal(input_shape)
>>> x2 = tf.random.normal(input_shape)
>>> y = tf.keras.layers.Add()([x1, x2])
>>> print(y.shape)
(2, 3, 4)
}\if{html}{\out{</div>}}

Used in a functional model:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> input1 = tf.keras.layers.Input(shape=(16,))
>>> x1 = tf.keras.layers.Dense(8, activation='relu')(input1)
>>> input2 = tf.keras.layers.Input(shape=(32,))
>>> x2 = tf.keras.layers.Dense(8, activation='relu')(input2)
>>> # equivalent to `added = tf.keras.layers.add([x1, x2])`
>>> added = tf.keras.layers.Add()([x1, x2])
>>> out = tf.keras.layers.Dense(4)(added)
>>> model = tf.keras.models.Model(inputs=[input1, input2], outputs=out)
}\if{html}{\out{</div>}}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
