% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_average}
\alias{layer_average}
\title{Layer that averages a list of inputs element-wise}
\usage{
layer_average(object, ...)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Layer that averages a list of inputs element-wise
}
\details{
It takes as input a list of tensors, all of the same shape, and returns
a single tensor (also of the same shape).
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
