% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_cropping_1d}
\alias{layer_cropping_1d}
\title{Cropping layer for 1D input (e.g. temporal sequence)}
\usage{
layer_cropping_1d(object, cropping = list(1L, 1L), ...)
}
\arguments{
\item{cropping}{Int or list of int (length 2)
How many units should be trimmed off at the beginning and end of
the cropping dimension (axis 1).
If a single int is provided, the same value will be used for both.}

\item{...}{standard layer arguments.}
}
\description{
It crops along the time dimension (axis 1).
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
