% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_gaussian_dropout}
\alias{layer_gaussian_dropout}
\title{Apply multiplicative 1-centered Gaussian noise}
\usage{
layer_gaussian_dropout(object, rate, seed = NULL, ...)
}
\arguments{
\item{rate}{Float, drop probability (as with \code{Dropout}).
The multiplicative noise will have
standard deviation \code{sqrt(rate / (1 - rate))}.}

\item{seed}{Integer, optional random seed to enable deterministic behavior.}

\item{...}{standard layer arguments.}
}
\description{
Apply multiplicative 1-centered Gaussian noise
}
\details{
As it is a regularization layer, it is only active at training time.

Call arguments:
inputs: Input tensor (of any rank).
training: Python boolean indicating whether the layer should behave in
training mode (adding dropout) or in inference mode (doing nothing).

Input shape:
Arbitrary. Use the keyword argument \code{input_shape}
(list of integers, does not include the samples axis)
when using this layer as the first layer in a model.

Output shape:
Same shape as input.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
