% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_global_average_pooling_1d}
\alias{layer_global_average_pooling_1d}
\title{Global average pooling operation for temporal data}
\usage{
layer_global_average_pooling_1d(object, data_format = "channels_last", ...)
}
\arguments{
\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch, steps, features)} while \code{channels_first}
corresponds to inputs with shape
\verb{(batch, features, steps)}.}

\item{...}{standard layer arguments.}

\item{keepdims}{A boolean, whether to keep the temporal dimension or not.
If \code{keepdims} is \code{FALSE} (default), the rank of the tensor is reduced
for spatial dimensions.
If \code{keepdims} is \code{TRUE}, the temporal dimension are retained with
length 1.
The behavior is the same as for \code{tf.reduce_mean} or \code{np.mean}.}
}
\description{
Global average pooling operation for temporal data
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
