% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_hashed_crossing}
\alias{layer_hashed_crossing}
\title{A preprocessing layer which crosses features using the "hashing trick"}
\usage{
layer_hashed_crossing(
  object,
  num_bins,
  output_mode = "int",
  sparse = FALSE,
  ...
)
}
\arguments{
\item{num_bins}{Number of hash bins.}

\item{output_mode}{Specification for the output of the layer. Values can be
\code{"int"}, or \code{"one_hot"} configuring the layer as follows:
\itemize{
\item \code{"int"}: Return the integer bin indices directly.
\item \code{"one_hot"}: Encodes each individual element in the input into an
array the same size as \code{num_bins}, containing a 1 at the input's
bin index. Defaults to \code{"int"}.
}}

\item{sparse}{Boolean. Only applicable to \code{"one_hot"} mode. If \code{TRUE},
returns a \code{SparseTensor} instead of a dense \code{Tensor}.
Defaults to \code{FALSE}.}

\item{...}{standard layer arguments.}

\item{**kwargs}{Keyword arguments to construct a layer.}
}
\description{
A preprocessing layer which crosses features using the "hashing trick"
}
\details{
This layer performs crosses of categorical features using the "hashing
trick". Conceptually, the transformation can be thought of as:
\verb{hash(concatenate(features)) \% num_bins}.

This layer currently only performs crosses of scalar inputs and batches of
scalar inputs. Valid input shapes are \verb{(batch_size, 1)}, \verb{(batch_size,)} and
\verb{()}.

For an overview and full list of preprocessing layers, see the preprocessing
\href{https://www.tensorflow.org/guide/keras/preprocessing_layers}{guide}.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
