% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_lambda}
\alias{layer_lambda}
\title{Wraps arbitrary expressions as a \code{Layer} object}
\usage{
layer_lambda(
  object,
  `function`,
  output_shape = NULL,
  mask = NULL,
  arguments = NULL,
  ...
)
}
\arguments{
\item{function}{The function to be evaluated. Takes input tensor as first
argument.}

\item{output_shape}{Expected output shape from function. This argument can be
inferred if not explicitly provided. Can be a list or function. If a
list, it only specifies the first dimension onward;
sample dimension is assumed either the same as the input:
\verb{output_shape = (input_shape[0], ) + output_shape} or, the input is
\code{NULL} and the sample dimension is also \code{NULL}:
\verb{output_shape = (NULL, ) + output_shape} If a function, it specifies the
entire shape as a function of the input shape:
\code{output_shape = f(input_shape)}}

\item{mask}{Either NULL (indicating no masking) or a callable with the same
signature as the \code{compute_mask} layer method, or a tensor that will be
returned as output mask regardless of what the input is.}

\item{arguments}{Optional dictionary of keyword arguments to be passed to the
function.}

\item{...}{standard layer arguments.}
}
\description{
Wraps arbitrary expressions as a \code{Layer} object
}
\details{
The \code{Lambda} layer exists so that arbitrary expressions can be used
as a \code{Layer} when constructing Sequential
and Functional API models. \code{Lambda} layers are best suited for simple
operations or quick experimentation. For more advanced use cases, follow
\href{https://www.tensorflow.org/guide/keras/custom_layers_and_models}{this guide}
for subclassing \code{tf.keras.layers.Layer}.

WARNING: \code{tf.keras.layers.Lambda} layers have (de)serialization limitations!

The main reason to subclass \code{tf.keras.layers.Layer} instead of using a
\code{Lambda} layer is saving and inspecting a Model. \code{Lambda} layers
are saved by serializing the Python bytecode, which is fundamentally
non-portable. They should only be loaded in the same environment where
they were saved. Subclassed layers can be saved in a more portable way
by overriding their \code{get_config()} method. Models that rely on
subclassed Layers are also often easier to visualize and reason about.

\if{html}{\out{<div class="sourceCode python">}}\preformatted{# add a x -> x^2 layer
model.add(Lambda(lambda x: x ** 2))
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode python">}}\preformatted{# add a layer that returns the concatenation
# of the positive part of the input and
# the opposite of the negative part

def antirectifier(x):
    x -= K.mean(x, axis=1, keepdims=TRUE)
    x = K.l2_normalize(x, axis=1)
    pos = K.relu(x)
    neg = K.relu(-x)
    return K.concatenate([pos, neg], axis=1)

model.add(Lambda(antirectifier))
}\if{html}{\out{</div>}}

\strong{Note on Variables:}

While it is possible to use Variables with Lambda layers,
this practice is discouraged as it can easily lead to bugs.
For instance, consider the following layer:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{scale = tf.Variable(1.)
scale_layer = tf.keras.layers.Lambda(lambda x: x * scale)
}\if{html}{\out{</div>}}

Because \code{scale_layer} does not directly track the \code{scale} variable, it will
not appear in \code{scale_layer.trainable_weights} and will therefore not be
trained if \code{scale_layer} is used in a Model.

A better pattern is to write a subclassed Layer:

\if{html}{\out{<div class="sourceCode python">}}\preformatted{class ScaleLayer(tf.keras.layers.Layer):
    def __init__(self, **kwargs):
        super().__init__(**kwargs)
        self.scale = tf.Variable(1.)

    def call(self, inputs):
        return inputs * self.scale
}\if{html}{\out{</div>}}

In general, \code{Lambda} layers can be convenient for simple stateless
computation, but anything more complex should use a subclass Layer instead.

Input shape: Arbitrary. Use the keyword argument input_shape (list of
integers, does not include the samples axis) when using this layer as the
first layer in a model.

Output shape: Specified by \code{output_shape} argument
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
