% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_permute}
\alias{layer_permute}
\title{Permutes the dimensions of the input according to a given pattern}
\usage{
layer_permute(object, dims, ...)
}
\arguments{
\item{dims}{Tuple of integers. Permutation pattern does not include the
samples dimension. Indexing starts at 1.
For instance, \verb{(2, 1)} permutes the first and second dimensions
of the input.}

\item{...}{standard layer arguments.}
}
\description{
Useful e.g. connecting RNNs and convnets.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
