% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_repeat_vector}
\alias{layer_repeat_vector}
\title{Repeats the input n times}
\usage{
layer_repeat_vector(object, n, ...)
}
\arguments{
\item{n}{Integer, repetition factor.}

\item{...}{standard layer arguments.}
}
\description{
Repeats the input n times
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
