% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_spatial_dropout_3d}
\alias{layer_spatial_dropout_3d}
\title{Spatial 3D version of Dropout}
\usage{
layer_spatial_dropout_3d(object, rate, data_format = NULL, ...)
}
\arguments{
\item{rate}{Float between 0 and 1. Fraction of the input units to drop.}

\item{data_format}{'channels_first' or 'channels_last'. In 'channels_first'
mode, the channels dimension (the depth) is at index 1, in
'channels_last' mode is it at index 4. When unspecified, uses
\code{image_data_format} value found in your Keras config file at
\verb{~/.keras/keras.json} (if exists) else 'channels_last'.
Defaults to 'channels_last'.}

\item{...}{standard layer arguments.}
}
\description{
Spatial 3D version of Dropout
}
\details{
This version performs the same function as Dropout, however, it drops
entire 3D feature maps instead of individual elements. If adjacent voxels
within feature maps are strongly correlated (as is normally the case in
early convolution layers) then regular dropout will not regularize the
activations and will otherwise just result in an effective learning rate
decrease. In this case, SpatialDropout3D will help promote independence
between feature maps and should be used instead.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
