% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_spectral_normalization}
\alias{layer_spectral_normalization}
\title{Performs spectral normalization on the weights of a target layer}
\usage{
layer_spectral_normalization(object, layer, power_iterations = 1L, ...)
}
\arguments{
\item{layer}{A \code{keras.layers.Layer} instance that
has either a \code{kernel} (e.g. \code{Conv2D}, \code{Dense}...)
or an \code{embeddings} attribute (\code{Embedding} layer).}

\item{power_iterations}{int, the number of iterations during normalization.}

\item{...}{standard layer arguments.}
}
\description{
Performs spectral normalization on the weights of a target layer
}
\details{
This wrapper controls the Lipschitz constant of the weights of a layer by
constraining their spectral norm, which can stabilize the training of GANs.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
