% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_stacked_rnn_cells}
\alias{layer_stacked_rnn_cells}
\title{Wrapper allowing a stack of RNN cells to behave as a single cell}
\usage{
layer_stacked_rnn_cells(cells, ...)
}
\arguments{
\item{cells}{List of RNN cell instances.}

\item{...}{standard layer arguments.}
}
\description{
Wrapper allowing a stack of RNN cells to behave as a single cell
}
\details{
Used to implement efficient stacked RNNs.

\if{html}{\out{<div class="sourceCode python">}}\preformatted{batch_size = 3
sentence_max_length = 5
n_features = 2
new_shape = (batch_size, sentence_max_length, n_features)
x = tf.constant(np.reshape(np.arange(30), new_shape), dtype = tf.float32)

rnn_cells = [tf.keras.layers.LSTMCell(128) for _ in range(2)]
stacked_lstm = tf.keras.layers.StackedRNNCells(rnn_cells)
lstm_layer = tf.keras.layers.RNN(stacked_lstm)

result = lstm_layer(x)
}\if{html}{\out{</div>}}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
