% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_subtract}
\alias{layer_subtract}
\title{Layer that subtracts two inputs}
\usage{
layer_subtract(inputs, ...)
}
\arguments{
\item{...}{standard layer arguments.}
}
\description{
Layer that subtracts two inputs
}
\details{
It takes as input a list of tensors of size 2, both of the same shape, and
returns a single tensor, (inputs\link{0} - inputs\link{1}), also of the same shape.

\if{html}{\out{<div class="sourceCode python">}}\preformatted{    import keras.src as keras

    input1 = keras.layers.Input(shape=(16,))
    x1 = keras.layers.Dense(8, activation='relu')(input1)
    input2 = keras.layers.Input(shape=(32,))
    x2 = keras.layers.Dense(8, activation='relu')(input2)
    # Equivalent to subtracted = keras.layers.subtract([x1, x2])
    subtracted = keras.layers.Subtract()([x1, x2])

    out = keras.layers.Dense(4)(subtracted)
    model = keras.models.Model(inputs=[input1, input2], outputs=out)
}\if{html}{\out{</div>}}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
