% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_upsampling_1d}
\alias{layer_upsampling_1d}
\title{Upsampling layer for 1D inputs}
\usage{
layer_upsampling_1d(object, size = 2L, ...)
}
\arguments{
\item{size}{Integer. Upsampling factor.}

\item{...}{standard layer arguments.}
}
\description{
Repeats each temporal step \code{size} times along the time axis.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
