% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_zero_padding_1d}
\alias{layer_zero_padding_1d}
\title{Zero-padding layer for 1D input (e.g. temporal sequence)}
\usage{
layer_zero_padding_1d(object, padding = 1L, ...)
}
\arguments{
\item{padding}{Int, or list of int (length 2).
\itemize{
\item If int:
How many zeros to add at the beginning and end of
the padding dimension (axis 1).
\item If list of int (length 2):
How many zeros to add at the beginning and the end of
the padding dimension (\verb{(left_pad, right_pad)}).
}}

\item{...}{standard layer arguments.}
}
\description{
Zero-padding layer for 1D input (e.g. temporal sequence)
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
