% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{layer_zero_padding_2d}
\alias{layer_zero_padding_2d}
\title{Zero-padding layer for 2D input (e.g. picture)}
\usage{
layer_zero_padding_2d(object, padding = list(1L, 1L), data_format = NULL, ...)
}
\arguments{
\item{padding}{Int, or list of 2 ints, or list of 2 lists of 2 ints.
\itemize{
\item If int: the same symmetric padding
is applied to height and width.
\item If list of 2 ints:
interpreted as two different
symmetric padding values for height and width:
\verb{(symmetric_height_pad, symmetric_width_pad)}.
\item If list of 2 lists of 2 ints:
interpreted as
\verb{((top_pad, bottom_pad), (left_pad, right_pad))}
}}

\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch_size, height, width, channels)} while \code{channels_first}
corresponds to inputs with shape
\verb{(batch_size, channels, height, width)}.
When unspecified, uses
\code{image_data_format} value found in your Keras config file at
\verb{~/.keras/keras.json} (if exists) else 'channels_last'.
Defaults to 'channels_last'.}

\item{...}{standard layer arguments.}
}
\description{
Zero-padding layer for 2D input (e.g. picture)
}
\details{
This layer can add rows and columns of zeros
at the top, bottom, left and right side of an image tensor.
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
