% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_zero_padding_3d}
\alias{layer_zero_padding_3d}
\title{Zero-padding layer for 3D data (spatial or spatio-temporal)}
\usage{
layer_zero_padding_3d(
  object,
  padding = list(1L, 1L, 1L),
  data_format = NULL,
  ...
)
}
\arguments{
\item{padding}{Int, or list of 3 ints, or list of 3 lists of 2 ints.
\itemize{
\item If int: the same symmetric padding
is applied to height and width.
\item If list of 3 ints:
interpreted as two different
symmetric padding values for height and width:
\verb{(symmetric_dim1_pad, symmetric_dim2_pad, symmetric_dim3_pad)}.
\item If list of 3 lists of 2 ints:
interpreted as
\verb{((left_dim1_pad, right_dim1_pad), (left_dim2_pad, right_dim2_pad), (left_dim3_pad, right_dim3_pad))}
}}

\item{data_format}{A string,
one of \code{channels_last} (default) or \code{channels_first}.
The ordering of the dimensions in the inputs.
\code{channels_last} corresponds to inputs with shape
\verb{(batch_size, spatial_dim1, spatial_dim2, spatial_dim3, channels)}
while \code{channels_first} corresponds to inputs with shape
\verb{(batch_size, channels, spatial_dim1, spatial_dim2, spatial_dim3)}.
When unspecified, uses
\code{image_data_format} value found in your Keras config file at
\verb{~/.keras/keras.json} (if exists) else 'channels_last'.
Defaults to 'channels_last'.}

\item{...}{standard layer arguments.}
}
\description{
Zero-padding layer for 3D data (spatial or spatio-temporal)
}
\details{
\if{html}{\out{<div class="sourceCode python">}}\preformatted{>>> input_shape = (1, 1, 2, 2, 3)
>>> x = np.arange(np.prod(input_shape)).reshape(input_shape)
>>> y = tf.keras.layers.ZeroPadding3D(padding=2)(x)
>>> print(y.shape)
(1, 5, 6, 6, 3)
}\if{html}{\out{</div>}}

Input shape:
5D tensor with shape:
\itemize{
\item If \code{data_format} is \code{"channels_last"}:
\verb{(batch_size, first_axis_to_pad, second_axis_to_pad, third_axis_to_pad, depth)}
\item If \code{data_format} is \code{"channels_first"}:
\verb{(batch_size, depth, first_axis_to_pad, second_axis_to_pad, third_axis_to_pad)}
}

Output shape:
5D tensor with shape:
\itemize{
\item If \code{data_format} is \code{"channels_last"}:
\verb{(batch_size, first_padded_axis, second_padded_axis, third_axis_to_pad, depth)}
\item If \code{data_format} is \code{"channels_first"}:
\verb{(batch_size, depth, first_padded_axis, second_padded_axis, third_axis_to_pad)}
}
}
\seealso{
\itemize{
\item \url{https://keras.io/api/layers}
}
}
